Find cells indicating shared biological features across conditions

Find cells indicating shared biological features across conditions

FindCell(seuratobj, seuratlist, fullcluster, distmat, firstn = 15)

Arguments

seuratobj

The Seurat object that all samples/subjects were merged together.

seuratlist

A list of Seurat objects, usually can be got by SplitObject().

fullcluster

A list of clusters that generated by the function GetCluster().

distmat

A list of distance vectors generated by the function FindNNDist().

firstn

The number of nearest cells were detected that you want to include in the permutation test. Default to be 15.

Value

A list of two vectors: one is for which cluster of which sample will be highlighted and the second one is which cells will be selected.

Examples

# Create example Seurat object
data(sim_data_sce)
data(sim_result)
# Create example list of Seurat objects
sim_data <- SCEtoSeurat(sim_data_sce)
#> Warning: No gene metadata columns found to add to Seurat object.
seuratlist <- Seurat::SplitObject(sim_data, split.by = "Study")
# seuratlist <- InputData(sim_data_sce,"Study")
# Create example fullcluster (mock data)
# fullcluster <- GetCluster(seuratlist)
# Create example distmat (mock data)
# distmat <- FindNNDist( fullcluster, distmat, meaningn = 20)

FindCell(sim_data, seuratlist, sim_result[[1]], sim_result[[3]], 15)
#> [[1]]
#> [1] "Sample2_1" "Sample2_2" "Sample2_3" "Sample2_4" "Sample3_4" "Sample3_5"
#> 
#> [[2]]
#>   [1] "TCTATGTGCGTGTA-1"   "AACTGTCTTCTAGG-1"   "CCTCTACTACCAAC-1"  
#>   [4] "CTTTAGACTGTCAG-1"   "ACTTGACTAACTGC-1"   "GCCAAATGTTGCAG-1"  
#>   [7] "AAGAAGACCTTGGA-1"   "GTGCAAACACTACG-1"   "ATTGCACTGGGACA-1"  
#>  [10] "AGAGGTCTCTTTAC-1"   "AAACATTGTTTGTC-1"   "CAATAATGGAAAGT-1"  
#>  [13] "GGAGACGAGCTGAT-1"   "CCATAGGACTTATC-1"   "ATGTCACTAACTGC-1"  
#>  [16] "AAGTAACTGATAGA-1"   "TCCATCCTACTGTG-1"   "GGAGGATGGGACTT-1"  
#>  [19] "TGCTGAGATGAGCT-1"   "TCATCATGGCGGAA-1"   "GACAACACCTCTTA-1"  
#>  [22] "CCAGCACTTAGACC-1"   "TATACAGACTCAGA-1"   "TCCACTCTACTAGC-1"  
#>  [25] "AGAACAGAGCTTAG-1"   "CGTCCAACGGTTTG-1"   "GTGGATTGTGCATG-1"  
#>  [28] "CCTCATCTGATGAA-1.1" "CAAGAAGACATACG-1"   "ACTTTGTGCCTCCA-1"  
#>  [31] "TTCTACGAGCGATT-1"   "AGGACACTCAGATC-1"   "ATTTCGTGAACAGA-1"  
#>  [34] "GATTCTACACCTGA-1"   "ATCGGAACTACGCA-1"   "ACGTTTACCTTAGG-1"  
#>  [37] "GCGAGCACCTGAAC-1"   "GGCGGACTTATCGG-1"   "ACTCAGGATGCTAG-1"  
#>  [40] "ATTGCTTGCAAGCT-1"   "CAAATTGAAGAACA-1"   "ACGATCGATGCGTA-1"  
#>  [43] "AATAGGGAGGTTAC-1"   "TGAACCGATGAGAA-1"   "CTTGAACTGAAACA-1"  
#>  [46] "ACAAATTGAACCAC-1"   "ATAGCCGATTTGGG-1"   "CTTGTATGGTTTCT-1"  
#>  [49] "GGACCCGACACTGA-1"   "GACTGAACCGTAAC-1"   "ATTGATGAAGAACA-1"  
#>  [52] "ATCTCAACAGAGTA-1"   "AGTTAAACTCCGTC-1"   "GAGCGCTGGGTCTA-1"  
#>  [55] "ACTCGAGACCTGTC-1"   "AGGACTTGTCGACA-1"   "TCAATCACCTAGAC-1"  
#>  [58] "GGAGGCCTTGAGAA-1"   "CACCGGGATGGCAT-1"   "AAGTTCCTAGATCC-1"  
#>  [61] "CGGTCACTTCACGA-1"   "CGCTACACACACGT-1"   "TGTTAAGAGTAGCT-1"  
#>  [64] "CCCATGTGACACGT-1"   "AGCGCCGATAGCCA-1"   "TACGCGCTCGCATA-1"  
#>  [67] "CTGAATCTGCTAAC-1"   "CTAACTACATTCTC-1"   "TTATGAGAGACGAG-1"  
#>  [70] "TAGGCTGAGGACGA-1"   "AGGTTGTGCCTAAG-1"   "ATTCAAGACCATGA-1"  
#>  [73] "TGGAGACTGTCTTT-1"   "TACGTACTGTCGAT-1"   "CGCAACCTAGGGTG-1"  
#>  [76] "GGCAAGGACCCTAC-1"   "TGCGAAACTCGACA-1"   "ATCACTTGTGTTTC-1"  
#>  [79] "GACACTGACGTAGT-1"   "ATTCGGGATCTTAC-1"   "CCACTTCTAAGTAG-1"  
#>  [82] "ACAGTGTGGGTGAG-1"   "GACAACTGGTCGTA-1"   "AAAGATCTGTTGCA-1"  
#>  [85] "AGCGCTCTGAGGCA-1"   "CCAGTGCTGGTGAG-1"   "ATGTCACTGGTACT-1"  
#>  [88] "ACAGTTCTGGACGA-1"   "GTATTCACCAGGAG-1"   "TGACACGACTGACA-1"  
#>  [91] "TACGGAACTGGCAT-1"   "GCTACGCTTGCCTC-1.1" "CCAGCACTCCATGA-1"  
#>  [94] "CGAGAACTGGAGGT-1"   "CATAGTCTGACGAG-1"   "GCAGTTGATCAAGC-1"  
#>  [97] "CGGTCACTTACAGC-1"   "CATTGACTGTATGC-1"   "CCACTGTGGCGAGA-1"  
#> [100] "TGTGACGAATGCCA-1"   "AGACGTACCTACGA-1"   "GTGTGATGGTTGAC-1"  
#> [103] "CCGGAGACAGAGGC-1"   "TAACCGGAACCCTC-1"   "GAAAGTGACGTACA-1"  
#> [106] "TTGAGGACTTGAGC-1"   "GGAGTTACTGCCAA-1"   "ACTACTACGTGTAC-1"  
#> [109] "CTTACATGTGGGAG-1"   "CGTGATGAATAAGG-1"   "CAGCTCACAACGAA-1"  
#> [112] "GCGCACGAGAGATA-1"   "CAAGGTTGAGATCC-1"   "TCCACGTGGGAGTG-1"  
#> [115] "GGGAAGACCGTAGT-1"   "ACAGACACTTTGGG-1"   "CGAGGCTGATTTCC-1"  
#> [118] "CCCGGAGATTGAGC-1"   "GACCATGACTGGAT-1"   "GAAGTGCTTCTCAT-1"  
#> [121] "GCGAGAGAAGCCAT-1"   "TCCATAACAAGGCG-1"   "CCAGTGCTGTGCTA-1"  
#> [124] "TAGAGCACCCTTTA-1"   "GATCATCTTTCCAT-1"   "CAGACATGACCTCC-1"  
#> [127] "ACCACAGACCTTGC-1"   "CGAGAACTCTGTAG-1"   "CATACTACAACGAA-1"  
#> [130] "TATCACTGGCATCA-1"   "CGTTAACTGGGAGT-1"   "AGTTTGCTATCACG-1"  
#> [133] "TATAAGTGGTGTCA-1"   "TACTACACTTGTCT-1"   "TTCTAGTGGGAAAT-1"  
#> [136] "CTCTAAACAGGAGC-1"   "ACCATTACAACCAC-1"   "ATGTAAACGAGCAG-1"  
#> [139] "AACATTGACATTCT-1"   "CAAGCATGGACAAA-1"   "TTTCACGAGGAACG-1"  
#> [142] "CCCTGAACGGAACG-1"   "AGAACAGAGGATTC-1"   "TGACTTACTTGCTT-1"  
#> [145] "TCGAGCCTGGGACA-1"   "GTGGATTGTCTTAC-1"   "TTAGGTCTTATGCG-1"  
#> [148] "AAGAAGACTGACTG-1"   "TCGAGCCTTGGTGT-1"   "AGCGATACGCGAGA-1"  
#> [151] "ATACGGACCATGGT-1"   "GTATTCACTGTCAG-1"   "ATATACGACCTCCA-1"  
#> [154] "TCTTCAGACAGTTG-1"   "CTCAGCTGGACACT-1"   "GGAGAGACCCTAAG-1"  
#> [157] "AAGATGGAACCACA-1"   "CATCCCGACACTCC-1"   "CATCCCGAGGATCT-1"  
#> [160] "TCAGTTACGGGAGT-1"   "GATAGCACATAAGG-1"   "CATTGTTGTCACCC-1"  
#> [163] "CTGGAAACGATAGA-1"   "CTCGACTGATTTCC-1"   "AACAATACCCCTAC-1"  
#> [166] "TTCCAAACGATGAA-1"   "CCAGCGGACTGAAC-1"   "GTGATCGAGTCTGA-1"  
#> [169] "TATTGCTGCGTAGT-1"   "TAGGAGCTCGTGTA-1"   "ACCCGTTGGCTTCC-1"  
#> [172] "AATAGGGAGTTCGA-1"   "GGAGGATGAGCTAC-1"   "TACTACTGTAAGGA-1"  
#> [175] "CTATACTGCCCTTG-1"   "CCAGCGGATCCTAT-1"   "TGAATAACGAATGA-1"  
#> [178] "CATCATACGGACAG-1"   "CCCTCAGAGAGACG-1"   "GGCGACACAAGTGA-1"  
#> [181] "AGCTTTACTTACCT-1"   "CTTCTAGAGCGAGA-1"   "ACTCCTCTCCACAA-1"  
#> [184] "CATGTTACGAATGA-1"   "CAGGAACTTGGGAG-1"   "CAATAAACGCTGTA-1"  
#> [187] "AATTGATGCTAGTG-1"   "AGCACAACGCCAAT-1"   "AGTTCTACATTCTC-1"  
#> [190] "GGGACCACTTTCAC-1"   "ACGCCGGACAGCTA-1"   "TTCCCACTTTGAGC-1"  
#> [193] "ACCTGAGAGTCTTT-1"   "GGCCGAACGTGCAT-1"   "TATACGCTCGTGAT-1"  
#> [196] "GTTGGATGTTGCGA-1"   "AAGGCTTGCCACAA-1"   "AGGGCCACAACCGT-1"  
#> [199] "CTATGACTGCAAGG-1"   "GCGTAAACCTCGAA-1"   "CTGAACGATACGCA-1"  
#> [202] "CAGTGATGGGGTGA-1"   "GCCACGGACATGAC-1"  
#>