Get broad and fine clusters

Get broad and fine clusters

GetCluster(seuratlist, n1 = 50, n2 = 200)

Arguments

seuratlist

A list of Seurat objects, usually can be got by SplitObject(). We also accept the SingleCellExperiment object input.

n1

If the number of cells was smaller than n1, then the cluster will remain unchanged called rare cluster. The default value of n1 is 50.

n2

If the count of cells within a broad cluster is more than n2, the cluster is subdivided randomly into three fine clusters. If the cell count falls within the range of n1 to n2, two fine clusters are generated randomly. Default value is 200.

Value

A list of data frames.

Examples

data(sim_data_sce)
# Assuming "seuratlist" is a list of Seurat objects
# if(is(sim_data_sce, "SingleCellExperiment")){ sim_data <- as.Seurat(sim_data_sce) }
sim_data <- SCEtoSeurat(sim_data_sce)
#> Warning: No gene metadata columns found to add to Seurat object.
seuratlist <- Seurat::SplitObject(sim_data, split.by = "Study")
# seuratlist <- InputData(sim_data_sce,"Study")
fullcluster <- GetCluster(seuratlist)
#> Normalizing layer: counts
#> Finding variable features for layer counts
#> Warning: pseudoinverse used at -2.5443
#> Warning: neighborhood radius 0.31902
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.090619
#> Centering and scaling data matrix
#> Warning: The following 994 features requested have zero variance; running reduction without them: ENSG00000172733	PURG, ENSG00000180209	MYLPF, ENSG00000261231	RP11-523L20.2, ENSG00000189367	KIAA0408, ENSG00000086570	FAT2, ENSG00000249882	RP11-665I14.1, ENSG00000237853	RP5-833A20.1, ENSG00000255276	RRM1-AS1, ENSG00000224902	GAGE12H, ENSG00000164185	ZNF474, ENSG00000248703	RP11-495K9.3, ENSG00000251151	HOXC-AS3, ENSG00000177076	ACER2, ENSG00000267038	RP11-1096D5.2, ENSG00000228569	AC073133.2, ENSG00000260780	RP11-580I1.1, ENSG00000231703	RP4-669H2.1, ENSG00000224658	RP11-631F7.1, ENSG00000250431	RP13-577H12.2, ENSG00000231731	AC010976.2, ENSG00000255420	RP11-236J17.3, ENSG00000204889	KRT40, ENSG00000237907	RP5-1013A22.2, ENSG00000185792	NLRP9, ENSG00000272579	RP11-101E13.5, ENSG00000272121	RP4-555D20.4, ENSG00000196754	S100A2, ENSG00000251532	CTD-2245E15.3, ENSG00000125954	CHURC1-FNTB, ENSG00000168621	GDNF, ENSG00000224844	AC107079.1, ENSG00000139988	RDH12, ENSG00000132693	CRP, ENSG00000174032	SLC25A30, ENSG00000257985	RP1-228P16.4, ENSG00000261583	CTD-2385L22.1, ENSG00000172260	NEGR1, ENSG00000230687	RP5-1114G22.2, ENSG00000261789	RP11-709D24.5, ENSG00000065320	NTN1, ENSG00000267467	APOC4, ENSG00000006788	MYH13, ENSG00000234622	RP13-297E16.5, ENSG00000233234	RP11-740P5.2, ENSG00000269064	AL160286.1, ENSG00000269895	AP000654.4, ENSG00000267580	CTD-2540B15.11, ENSG00000265778	RP11-17M16.2, ENSG00000269199	AC003973.5, ENSG00000184492	FOXD4L1, ENSG00000249847	RP11-192C21.2, ENSG00000226113	RP11-672F9.1, ENSG00000255104	ZNF286A, ENSG00000263508	RP11-963H4.3, ENSG00000231294	AC006994.2, ENSG00000064309	CDON, ENSG00000254290	RP11-150O12.3, ENSG00000261717	RP11-77K12.1, ENSG00000184454	NCMAP, ENSG00000228539	RP11-526K17.2, ENSG00000236306	RP11-33G16.1, ENSG00000099364	FBXL19, ENSG00000204187	LINC00619, ENSG00000120436	GPR31, ENSG00000227131	RP11-552E20.1, ENSG00000245385	RP11-334E6.10, ENSG00000133131	MORC4, ENSG00000242767	ZBTB20-AS4, ENSG00000226386	PARD3-AS1, ENSG00000165309	ARMC3, ENSG00000170835	CEL, ENSG00000256391	SDIM1, ENSG00000232530	RP1-102K2.6, ENSG00000235910	APOA1-AS, ENSG00000185894	BPY2C, ENSG00000198400	NTRK1, ENSG00000258888	RP11-326A13.1, ENSG00000235085	AC091153.4, ENSG00000204702	OR2J1, ENSG00000165996	PTPLA, ENSG00000163251	FZD5, ENSG00000259818	RP5-1024G6.7, ENSG00000203301	AL590822.1, ENSG00000178412	RP11-567M16.3, ENSG00000233835	RP1-92C4.2, ENSG00000268012	AC016559.1, ENSG00000268549	RP11-295P9.8, ENSG00000175121	WFDC5, ENSG00000259063	DUX4L5, ENSG00000228213	NLGN1-AS1, ENSG00000233821	ENOX1-AS1, ENSG00000221938	OR2A14, ENSG00000091128	LAMB4, ENSG00000224691	GS1-174L6.4, ENSG00000183032	SLC25A21, ENSG00000259312	RP11-522B15.5, ENSG00000270755	RP11-138E2.1, ENSG00000145198	VWA5B2, ENSG00000174111	SOCS7, ENSG00000100012	SEC14L3, ENSG00000181074	OR52N4, ENSG00000250241	RP11-9G1.3, ENSG00000236761	CTAGE9, ENSG00000248828	RP11-673E1.4, ENSG00000204110	RP1-153P14.8, ENSG00000227518	MIR1302-9, ENSG00000224201	PNMA6A, ENSG00000225216	AC007362.3, ENSG00000251258	RFPL4B, ENSG00000249639	CTB-138E5.1, ENSG00000156486	KCNS2, ENSG00000228272	AC106874.1, ENSG00000163914	RHO, ENSG00000196532	HIST1H3C, ENSG00000256802	RP11-680F8.1, ENSG00000134853	PDGFRA, ENSG00000117479	SLC19A2, ENSG00000228214	LINC00693, ENSG00000205537	RP11-89H19.1, ENSG00000272663	RP11-191L17.1, ENSG00000229855	CTC-546K23.1, ENSG00000236463	LINC00427, ENSG00000259044	RP11-789A21.1, ENSG00000268091	AC078925.1, ENSG00000254768	CTD-2140G10.2, ENSG00000125257	ABCC4, ENSG00000268126	AC091948.1, ENSG00000272762	RP11-155D18.14, ENSG00000164236	ANKRD33B, ENSG00000123243	ITIH5, ENSG00000165066	NKX6-3, ENSG00000172382	PRSS27, ENSG00000232529	RP1-257I9.2, ENSG00000231734	RP6-206I17.2, ENSG00000222030	AC007131.1, ENSG00000203364	RP11-370F5.4, ENSG00000240553	RP1-184J9.2, ENSG00000099812	MISP, ENSG00000253893	FAM85B, ENSG00000176742	OR51V1, ENSG00000233244	CYP4F59P, ENSG00000259518	RP11-597K23.2, ENSG00000069764	PLA2G10, ENSG00000253735	RP11-115J16.3, ENSG00000230920	RP11-527D15.1, ENSG00000262576	PCDHGA4, ENSG00000130433	CACNG6, ENSG00000140993	TIGD7, ENSG00000004776	HSPB6, ENSG00000205693	MANSC4, ENSG00000273091	AP000255.6, ENSG00000143797	MBOAT2, ENSG00000170549	IRX1, ENSG00000261272	MUC22, ENSG00000227782	AC002553.1, ENSG00000115602	IL1RL1, ENSG00000176754	LINC00303, ENSG00000131480	AOC2, ENSG00000140798	ABCC12, ENSG00000164855	TMEM184A, ENSG00000233670	PIRT, ENSG00000007171	NOS2, ENSG00000226957	RP4-533D7.4, ENSG00000187730	GABRD, ENSG00000259375	RP11-815J21.2, ENSG00000255327	RP11-555G19.1, ENSG00000099377	HSD3B7, ENSG00000267769	CTB-50L17.9, ENSG00000157557	ETS2, ENSG00000227047	AC012065.4, ENSG00000206043	C18orf63, ENSG00000165985	C1QL3, ENSG00000231394	RP11-310H4.3, ENSG00000124334	IL9R, ENSG00000226209	AL138764.1, ENSG00000267191	RP11-15A1.2, ENSG00000198812	LRRC10, ENSG00000255313	RP11-1082L8.2, ENSG00000268809	AC012360.2, ENSG00000231426	RP5-899B16.1, ENSG00000224616	RP11-305E17.6, ENSG00000259714	LINC00594, ENSG00000254084	KB-1930G5.4, ENSG00000246331	RP11-77I22.2, ENSG00000130701	RBBP8NL, ENSG00000269676	MUC8, ENSG00000268923	AL645922.1, ENSG00000139797	RNF113B, ENSG00000171540	OTP, ENSG00000257918	RP11-482D24.3, ENSG00000256209	RP11-897M7.1, ENSG00000257894	RP1-78O14.1, ENSG00000182578	CSF1R, ENSG00000264750	RP11-277J6.2, ENSG00000197471	SPN, ENSG00000130822	PNCK, ENSG00000058085	LAMC2, ENSG00000243224	RP5-1157M23.2, ENSG00000167858	TEKT1, ENSG00000124813	RUNX2, ENSG00000260235	CTD-2026K11.3, ENSG00000271973	RP11-572O6.1, ENSG00000232080	XXbac-BPG254F23.7, ENSG00000231171	LINC01098, ENSG00000226310	RP3-323P24.3, ENSG00000225872	AC002398.5, ENSG00000260201	RP11-143N13.2, ENSG00000260725	AC005307.1, ENSG00000258223	PRSS58, ENSG00000053524	MCF2L2, ENSG00000174992	ZG16, ENSG00000235631	RNF148, ENSG00000244300	RP11-475N22.4, ENSG00000260464	RP4-561L24.3, ENSG00000268107	AC003005.4, ENSG00000258810	RP11-219E7.1, ENSG00000173431	RNASE8, ENSG00000174807	CD248, ENSG00000267259	CTD-2008P7.9, ENSG00000273281	RP11-339B21.14, ENSG00000170921	TANC2, ENSG00000229084	LINC00226, ENSG00000253644	KB-1930G5.3, ENSG00000272656	RP11-219D15.3, ENSG00000121742	GJB6, ENSG00000253392	AC006277.2, ENSG00000238004	AC009227.3, ENSG00000267607	CTD-2369P2.8, ENSG00000255322	RP11-22P4.1, ENSG00000205622	AF064858.6, ENSG00000151388	ADAMTS12, ENSG00000224739	AC016735.1, ENSG00000242104	RP11-340E6.1, ENSG00000250337	LINC01021, ENSG00000259437	RP11-798K3.4, ENSG00000228526	RP3-510D11.1, ENSG00000162711	NLRP3, ENSG00000155719	OTOA, ENSG00000257787	RP11-20L19.1, ENSG00000160505	NLRP4, ENSG00000235586	AC011247.3, ENSG00000258926	RP11-47I22.1, ENSG00000126785	RHOJ, ENSG00000159556	ISL2, ENSG00000253524	RP11-89M20.2, ENSG00000272817	RP11-402D21.2, ENSG00000225265	RP11-378J18.3, ENSG00000258867	LINC01146, ENSG00000168246	UBTD2, ENSG00000259364	RP11-64K12.9, ENSG00000168903	BTNL3, ENSG00000179941	BBS10, ENSG00000100068	LRP5L, ENSG00000250234	CTD-2024P10.1, ENSG00000232183	RP6-27P15.2, ENSG00000258455	RP11-665C16.5, ENSG00000267552	CTD-2528L19.4, ENSG00000120094	HOXB1, ENSG00000176358	TAC4, ENSG00000273104	CMP21-97G8.1, ENSG00000254902	ANO1-AS1, ENSG00000101435	CST9L, ENSG00000254187	CTB-78F1.1, ENSG00000248864	RP11-348J24.1, ENSG00000184985	SORCS2, ENSG00000204481	PRAMEF14, ENSG00000235772	AP001625.6, ENSG00000267070	RP11-10K17.6, ENSG00000272040	CTC-366B18.4, ENSG00000158859	ADAMTS4, ENSG00000234551	RP11-123H22.1, ENSG00000231297	SNORD45, ENSG00000118972	FGF23, ENSG00000229272	RP11-498J9.2, ENSG00000177822	AC108142.1, ENSG00000106299	WASL, ENSG00000133115	STOML3, ENSG00000236544	AC008060.8, ENSG00000134871	COL4A2, ENSG00000166869	CHP2, ENSG00000173401	GLIPR1L1, ENSG00000162757	C1orf74, ENSG00000225194	LINC00092, ENSG00000189266	PNRC2, ENSG00000261194	RP11-704C2.1, ENSG00000255355	AP000640.2, ENSG00000267838	AC008746.12, ENSG00000215549	AL138815.1, ENSG00000187612	OR5W2, ENSG00000238042	RP11-815M8.1, ENSG00000269709	AL109659.1, ENSG00000177476	OR2G3, ENSG00000188859	FAM78B, ENSG00000228302	RP11-186N15.3, ENSG00000272962	SLC5A3, ENSG00000257523	CTD-2384A14.2, ENSG00000237189	RP11-85G21.2, ENSG00000261020	RP11-744K17.1, ENSG00000226516	FAM138B, ENSG00000260977	RP11-333I13.1, ENSG00000241732	RP11-38P22.2, ENSG0000025
#> PC_ 1 
#> Positive:  ENSG00000122026	RPL21, ENSG00000063177	RPL18, ENSG00000162244	RPL29, ENSG00000105373	GLTSCR2, ENSG00000105193	RPS16, ENSG00000165983	PTER, ENSG00000145416	MARCH1, ENSG00000073614	KDM5A, ENSG00000109787	KLF3, ENSG00000167483	FAM129C 
#> 	   ENSG00000115947	ORC4, ENSG00000084234	APLP2, ENSG00000130787	HIP1R, ENSG00000262202	SNORD3D, ENSG00000175155	YPEL2, ENSG00000267680	ZNF224, ENSG00000271858	XXcos-LUCA11.4, ENSG00000214900	C14orf182, ENSG00000163728	TTC14, ENSG00000145439	CBR4 
#> 	   ENSG00000272269	RP11-500C11.3, ENSG00000204482	LST1, ENSG00000172216	CEBPB, ENSG00000186166	CCDC84, ENSG00000118518	RNF146, ENSG00000085982	USP40, ENSG00000257698	RP11-620J15.3, ENSG00000173933	RBM4, ENSG00000204252	HLA-DOA, ENSG00000157106	SMG1 
#> Negative:  ENSG00000166387	PPFIBP2, ENSG00000124787	RPP40, ENSG00000082641	NFE2L1, ENSG00000143228	NUF2, ENSG00000253426	RP11-10A14.4, ENSG00000152253	SPC25, ENSG00000167513	CDT1, ENSG00000072571	HMMR, ENSG00000169679	BUB1, ENSG00000111206	FOXM1 
#> 	   ENSG00000157456	CCNB2, ENSG00000155324	GRAMD3, ENSG00000257303	RP11-977G19.11, ENSG00000163467	TSACC, ENSG00000186376	ZNF75D, ENSG00000102575	ACP5, ENSG00000123240	OPTN, ENSG00000139193	CD27, ENSG00000125817	CENPB, ENSG00000134014	ELP3 
#> 	   ENSG00000106683	LIMK1, ENSG00000173457	PPP1R14B, ENSG00000197157	SND1, ENSG00000108010	GLRX3, ENSG00000213977	TAX1BP3, ENSG00000260083	MIR4519, ENSG00000167136	ENDOG, ENSG00000030582	GRN, ENSG00000196505	GDAP2, ENSG00000158481	CD1C 
#> PC_ 2 
#> Positive:  ENSG00000223865	HLA-DPB1, ENSG00000158481	CD1C, ENSG00000239961	LILRA4, ENSG00000112799	LY86, ENSG00000099956	SMARCB1, ENSG00000166228	PCBD1, ENSG00000161091	MFSD12, ENSG00000172366	FAM195A, ENSG00000125726	CD70, ENSG00000163156	SCNM1 
#> 	   ENSG00000148719	DNAJB12, ENSG00000012779	ALOX5, ENSG00000173559	NABP1, ENSG00000127481	UBR4, ENSG00000072694	FCGR2B, ENSG00000107581	EIF3A, ENSG00000213240	NOTCH2NL, ENSG00000102575	ACP5, ENSG00000066336	SPI1, ENSG00000175893	ZDHHC21 
#> 	   ENSG00000099330	OCEL1, ENSG00000108588	CCDC47, ENSG00000129534	MIS18BP1, ENSG00000134884	ARGLU1, ENSG00000080189	SLC35C2, ENSG00000076685	NT5C2, ENSG00000112159	MDN1, ENSG00000185347	C14orf80, ENSG00000174004	NRROS, ENSG00000155657	TTN 
#> Negative:  ENSG00000090554	FLT3LG, ENSG00000122026	RPL21, ENSG00000168209	DDIT4, ENSG00000125817	CENPB, ENSG00000197880	MDS2, ENSG00000082641	NFE2L1, ENSG00000198821	CD247, ENSG00000154764	WNT7A, ENSG00000253426	RP11-10A14.4, ENSG00000168496	FEN1 
#> 	   ENSG00000163467	TSACC, ENSG00000153130	SCOC, ENSG00000160087	UBE2J2, ENSG00000066739	ATG2B, ENSG00000027075	PRKCH, ENSG00000148950	IMMP1L, ENSG00000143228	NUF2, ENSG00000111206	FOXM1, ENSG00000152253	SPC25, ENSG00000072571	HMMR 
#> 	   ENSG00000155324	GRAMD3, ENSG00000167513	CDT1, ENSG00000157456	CCNB2, ENSG00000169679	BUB1, ENSG00000186376	ZNF75D, ENSG00000257303	RP11-977G19.11, ENSG00000122971	ACADS, ENSG00000238227	C9orf69, ENSG00000120915	EPHX2, ENSG00000172216	CEBPB 
#> PC_ 3 
#> Positive:  ENSG00000198821	CD247, ENSG00000139193	CD27, ENSG00000090554	FLT3LG, ENSG00000159496	RGL4, ENSG00000168209	DDIT4, ENSG00000166971	AKTIP, ENSG00000027075	PRKCH, ENSG00000101474	APMAP, ENSG00000129932	DOHH, ENSG00000173221	GLRX 
#> 	   ENSG00000111348	ARHGDIB, ENSG00000137193	PIM1, ENSG00000187118	CMC1, ENSG00000180096	SEPT1, ENSG00000167815	PRDX2, ENSG00000172757	CFL1, ENSG00000160888	IER2, ENSG00000141522	ARHGDIA, ENSG00000099330	OCEL1, ENSG00000235999	RP11-403I13.8 
#> 	   ENSG00000026025	VIM, ENSG00000109046	WSB1, ENSG00000109062	SLC9A3R1, ENSG00000167118	URM1, ENSG00000118680	MYL12B, ENSG00000120915	EPHX2, ENSG00000108010	GLRX3, ENSG00000141101	NOB1, ENSG00000273080	RP11-301O19.1, ENSG00000100129	EIF3L 
#> Negative:  ENSG00000223865	HLA-DPB1, ENSG00000257303	RP11-977G19.11, ENSG00000100908	EMC9, ENSG00000186376	ZNF75D, ENSG00000143228	NUF2, ENSG00000253426	RP11-10A14.4, ENSG00000163467	TSACC, ENSG00000082641	NFE2L1, ENSG00000173457	PPP1R14B, ENSG00000152253	SPC25 
#> 	   ENSG00000167513	CDT1, ENSG00000169679	BUB1, ENSG00000111206	FOXM1, ENSG00000157456	CCNB2, ENSG00000072571	HMMR, ENSG00000155324	GRAMD3, ENSG00000196505	GDAP2, ENSG00000148840	PPRC1, ENSG00000128039	SRD5A3, ENSG00000111224	PARP11 
#> 	   ENSG00000112799	LY86, ENSG00000156313	RPGR, ENSG00000166387	PPFIBP2, ENSG00000165733	BMS1, ENSG00000227039	ITGB2-AS1, ENSG00000126391	FRMD8, ENSG00000247240	UBL7-AS1, ENSG00000160741	CRTC2, ENSG00000106683	LIMK1, ENSG00000205155	PSENEN 
#> PC_ 4 
#> Positive:  ENSG00000170175	CHRNB1, ENSG00000126368	NR1D1, ENSG00000112640	PPP2R5D, ENSG00000154764	WNT7A, ENSG00000054148	PHPT1, ENSG00000105329	TGFB1, ENSG00000120915	EPHX2, ENSG00000134460	IL2RA, ENSG00000110330	BIRC2, ENSG00000153201	RANBP2 
#> 	   ENSG00000020256	ZFP64, ENSG00000120948	TARDBP, ENSG00000163564	PYHIN1, ENSG00000127527	EPS15L1, ENSG00000168806	LCMT2, ENSG00000148218	ALAD, ENSG00000173442	EHBP1L1, ENSG00000100083	GGA1, ENSG00000072364	AFF4, ENSG00000158481	CD1C 
#> 	   ENSG00000136819	C9orf78, ENSG00000189339	SLC35E2B, ENSG00000198498	TMA16, ENSG00000150776	C11orf57, ENSG00000270170	NCBP2-AS2, ENSG00000150054	MPP7, ENSG00000178773	CPNE7, ENSG00000088766	CRLS1, ENSG00000173559	NABP1, ENSG00000173264	GPR137 
#> Negative:  ENSG00000149929	HIRIP3, ENSG00000233937	CTC-338M12.4, ENSG00000153896	ZNF599, ENSG00000167208	SNX20, ENSG00000135913	USP37, ENSG00000237438	CECR7, ENSG00000004468	CD38, ENSG00000095319	NUP188, ENSG00000271856	RP11-861A13.4, ENSG00000121064	SCPEP1 
#> 	   ENSG00000225828	FAM229A, ENSG00000273080	RP11-301O19.1, ENSG00000079246	XRCC5, ENSG00000196642	RABL6, ENSG00000168818	STX18, ENSG00000196230	TUBB, ENSG00000124357	NAGK, ENSG00000159753	RLTPR, ENSG00000112159	MDN1, ENSG00000180957	PITPNB 
#> 	   ENSG00000197989	SNHG12, ENSG00000003249	DBNDD1, ENSG00000196262	PPIA, ENSG00000080189	SLC35C2, ENSG00000064651	SLC12A2, ENSG00000132424	PNISR, ENSG00000128191	DGCR8, ENSG00000198355	PIM3, ENSG00000211456	SACM1L, ENSG00000100416	TRMU 
#> PC_ 5 
#> Positive:  ENSG00000173221	GLRX, ENSG00000167196	FBXO22, ENSG00000125995	ROMO1, ENSG00000109534	GAR1, ENSG00000221916	C19orf73, ENSG00000109118	PHF12, ENSG00000167118	URM1, ENSG00000106355	LSM5, ENSG00000139921	TMX1, ENSG00000173559	NABP1 
#> 	   ENSG00000241990	RP6-109B7.3, ENSG00000169905	TOR1AIP2, ENSG00000150776	C11orf57, ENSG00000129932	DOHH, ENSG00000152127	MGAT5, ENSG00000180423	HARBI1, ENSG00000108010	GLRX3, ENSG00000219545	RPA3-AS1, ENSG00000174456	C12orf76, ENSG00000131196	NFATC1 
#> 	   ENSG00000156515	HK1, ENSG00000163960	UBXN7, ENSG00000162642	C1orf52, ENSG00000108774	RAB5C, ENSG00000163156	SCNM1, ENSG00000134109	EDEM1, ENSG00000101220	C20orf27, ENSG00000127838	PNKD, ENSG00000236104	ZBTB22, ENSG00000142856	ITGB3BP 
#> Negative:  ENSG00000229539	RP11-119B16.2, ENSG00000126368	NR1D1, ENSG00000148218	ALAD, ENSG00000145545	SRD5A1, ENSG00000120697	ALG5, ENSG00000109046	WSB1, ENSG00000168818	STX18, ENSG00000112640	PPP2R5D, ENSG00000112096	SOD2, ENSG00000141101	NOB1 
#> 	   ENSG00000118900	UBN1, ENSG00000154764	WNT7A, ENSG00000117528	ABCD3, ENSG00000114857	NKTR, ENSG00000198546	ZNF511, ENSG00000171302	CANT1, ENSG00000204954	C12orf73, ENSG00000108826	MRPL27, ENSG00000015479	MATR3, ENSG00000125637	PSD4 
#> 	   ENSG00000128524	ATP6V1F, ENSG00000137404	NRM, ENSG00000132549	VPS13B, ENSG00000174886	NDUFA11, ENSG00000115128	SF3B14, ENSG00000153814	JAZF1, ENSG00000079246	XRCC5, ENSG00000139197	PEX5, ENSG00000054148	PHPT1, ENSG00000096968	JAK2 
#> Computing nearest neighbor graph
#> Computing SNN
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 336
#> Number of edges: 21610
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.5099
#> Number of communities: 2
#> Elapsed time: 0 seconds
#> Normalizing layer: counts
#> Finding variable features for layer counts
#> Warning: pseudoinverse used at -2.1491
#> Warning: neighborhood radius 0.3166
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.031008
#> Centering and scaling data matrix
#> Warning: The following 1117 features requested have zero variance; running reduction without them: ENSG00000172733	PURG, ENSG00000174529	TMEM81, ENSG00000180209	MYLPF, ENSG00000261231	RP11-523L20.2, ENSG00000189367	KIAA0408, ENSG00000086570	FAT2, ENSG00000249882	RP11-665I14.1, ENSG00000237853	RP5-833A20.1, ENSG00000228857	AC104653.1, ENSG00000255276	RRM1-AS1, ENSG00000224902	GAGE12H, ENSG00000164185	ZNF474, ENSG00000248703	RP11-495K9.3, ENSG00000251151	HOXC-AS3, ENSG00000177076	ACER2, ENSG00000134014	ELP3, ENSG00000267038	RP11-1096D5.2, ENSG00000228569	AC073133.2, ENSG00000260780	RP11-580I1.1, ENSG00000231703	RP4-669H2.1, ENSG00000224658	RP11-631F7.1, ENSG00000250431	RP13-577H12.2, ENSG00000169398	PTK2, ENSG00000231731	AC010976.2, ENSG00000255420	RP11-236J17.3, ENSG00000204889	KRT40, ENSG00000237907	RP5-1013A22.2, ENSG00000185792	NLRP9, ENSG00000272579	RP11-101E13.5, ENSG00000272121	RP4-555D20.4, ENSG00000196754	S100A2, ENSG00000169679	BUB1, ENSG00000251532	CTD-2245E15.3, ENSG00000125954	CHURC1-FNTB, ENSG00000168621	GDNF, ENSG00000224844	AC107079.1, ENSG00000139988	RDH12, ENSG00000132693	CRP, ENSG00000174032	SLC25A30, ENSG00000257985	RP1-228P16.4, ENSG00000261583	CTD-2385L22.1, ENSG00000172260	NEGR1, ENSG00000230687	RP5-1114G22.2, ENSG00000261789	RP11-709D24.5, ENSG00000065320	NTN1, ENSG00000267467	APOC4, ENSG00000006788	MYH13, ENSG00000234622	RP13-297E16.5, ENSG00000233234	RP11-740P5.2, ENSG00000269064	AL160286.1, ENSG00000269895	AP000654.4, ENSG00000267580	CTD-2540B15.11, ENSG00000265778	RP11-17M16.2, ENSG00000269199	AC003973.5, ENSG00000184492	FOXD4L1, ENSG00000185305	ARL15, ENSG00000249847	RP11-192C21.2, ENSG00000226113	RP11-672F9.1, ENSG00000255104	ZNF286A, ENSG00000214960	ISPD, ENSG00000230424	RP1-43E13.2, ENSG00000064309	CDON, ENSG00000254290	RP11-150O12.3, ENSG00000104490	NCALD, ENSG00000261717	RP11-77K12.1, ENSG00000184454	NCMAP, ENSG00000228539	RP11-526K17.2, ENSG00000255062	RP11-712L6.5, ENSG00000236306	RP11-33G16.1, ENSG00000099364	FBXL19, ENSG00000204187	LINC00619, ENSG00000120436	GPR31, ENSG00000178199	ZC3H12D, ENSG00000227131	RP11-552E20.1, ENSG00000245385	RP11-334E6.10, ENSG00000133131	MORC4, ENSG00000175322	ZNF519, ENSG00000242767	ZBTB20-AS4, ENSG00000074696	PTPLAD1, ENSG00000226386	PARD3-AS1, ENSG00000165309	ARMC3, ENSG00000170835	CEL, ENSG00000256391	SDIM1, ENSG00000232530	RP1-102K2.6, ENSG00000115904	SOS1, ENSG00000235910	APOA1-AS, ENSG00000185894	BPY2C, ENSG00000198400	NTRK1, ENSG00000258888	RP11-326A13.1, ENSG00000235085	AC091153.4, ENSG00000198799	LRIG2, ENSG00000204702	OR2J1, ENSG00000156313	RPGR, ENSG00000197937	ZNF347, ENSG00000165996	PTPLA, ENSG00000203301	AL590822.1, ENSG00000178412	RP11-567M16.3, ENSG00000233835	RP1-92C4.2, ENSG00000268012	AC016559.1, ENSG00000268549	RP11-295P9.8, ENSG00000175121	WFDC5, ENSG00000259063	DUX4L5, ENSG00000228213	NLGN1-AS1, ENSG00000233821	ENOX1-AS1, ENSG00000139514	SLC7A1, ENSG00000221938	OR2A14, ENSG00000091128	LAMB4, ENSG00000224691	GS1-174L6.4, ENSG00000183032	SLC25A21, ENSG00000259312	RP11-522B15.5, ENSG00000263335	AF001548.5, ENSG00000270755	RP11-138E2.1, ENSG00000145198	VWA5B2, ENSG00000174111	SOCS7, ENSG00000100012	SEC14L3, ENSG00000181074	OR52N4, ENSG00000250241	RP11-9G1.3, ENSG00000236761	CTAGE9, ENSG00000103528	SYT17, ENSG00000248828	RP11-673E1.4, ENSG00000077458	FAM76B, ENSG00000204110	RP1-153P14.8, ENSG00000227518	MIR1302-9, ENSG00000224201	PNMA6A, ENSG00000225216	AC007362.3, ENSG00000251258	RFPL4B, ENSG00000249639	CTB-138E5.1, ENSG00000156486	KCNS2, ENSG00000228272	AC106874.1, ENSG00000163914	RHO, ENSG00000196532	HIST1H3C, ENSG00000256802	RP11-680F8.1, ENSG00000134853	PDGFRA, ENSG00000117479	SLC19A2, ENSG00000228214	LINC00693, ENSG00000205537	RP11-89H19.1, ENSG00000272663	RP11-191L17.1, ENSG00000229855	CTC-546K23.1, ENSG00000236463	LINC00427, ENSG00000238058	RP11-432J22.2, ENSG00000259044	RP11-789A21.1, ENSG00000268091	AC078925.1, ENSG00000254768	CTD-2140G10.2, ENSG00000130684	ZNF337, ENSG00000125257	ABCC4, ENSG00000268126	AC091948.1, ENSG00000272762	RP11-155D18.14, ENSG00000164236	ANKRD33B, ENSG00000123243	ITIH5, ENSG00000165066	NKX6-3, ENSG00000172382	PRSS27, ENSG00000232529	RP1-257I9.2, ENSG00000231734	RP6-206I17.2, ENSG00000222030	AC007131.1, ENSG00000203364	RP11-370F5.4, ENSG00000240553	RP1-184J9.2, ENSG00000099812	MISP, ENSG00000253893	FAM85B, ENSG00000176742	OR51V1, ENSG00000233244	CYP4F59P, ENSG00000259518	RP11-597K23.2, ENSG00000069764	PLA2G10, ENSG00000253735	RP11-115J16.3, ENSG00000230920	RP11-527D15.1, ENSG00000262576	PCDHGA4, ENSG00000130433	CACNG6, ENSG00000140993	TIGD7, ENSG00000004776	HSPB6, ENSG00000205693	MANSC4, ENSG00000273091	AP000255.6, ENSG00000170549	IRX1, ENSG00000261272	MUC22, ENSG00000172046	USP19, ENSG00000227782	AC002553.1, ENSG00000115602	IL1RL1, ENSG00000176754	LINC00303, ENSG00000131480	AOC2, ENSG00000140798	ABCC12, ENSG00000164855	TMEM184A, ENSG00000233670	PIRT, ENSG00000095319	NUP188, ENSG00000007171	NOS2, ENSG00000226957	RP4-533D7.4, ENSG00000152969	JAKMIP1, ENSG00000187730	GABRD, ENSG00000255327	RP11-555G19.1, ENSG00000099377	HSD3B7, ENSG00000267769	CTB-50L17.9, ENSG00000227047	AC012065.4, ENSG00000206043	C18orf63, ENSG00000165985	C1QL3, ENSG00000231394	RP11-310H4.3, ENSG00000226209	AL138764.1, ENSG00000267191	RP11-15A1.2, ENSG00000198812	LRRC10, ENSG00000186376	ZNF75D, ENSG00000255313	RP11-1082L8.2, ENSG00000268809	AC012360.2, ENSG00000231426	RP5-899B16.1, ENSG00000224616	RP11-305E17.6, ENSG00000259714	LINC00594, ENSG00000254084	KB-1930G5.4, ENSG00000246331	RP11-77I22.2, ENSG00000130701	RBBP8NL, ENSG00000269676	MUC8, ENSG00000268472	AP002884.2, ENSG00000268923	AL645922.1, ENSG00000139797	RNF113B, ENSG00000171540	OTP, ENSG00000257918	RP11-482D24.3, ENSG00000256209	RP11-897M7.1, ENSG00000257894	RP1-78O14.1, ENSG00000264750	RP11-277J6.2, ENSG00000228696	ARL17B, ENSG00000130822	PNCK, ENSG00000058085	LAMC2, ENSG00000243224	RP5-1157M23.2, ENSG00000167858	TEKT1, ENSG00000260235	CTD-2026K11.3, ENSG00000271973	RP11-572O6.1, ENSG00000232080	XXbac-BPG254F23.7, ENSG00000231171	LINC01098, ENSG00000226310	RP3-323P24.3, ENSG00000225872	AC002398.5, ENSG00000260201	RP11-143N13.2, ENSG00000260725	AC005307.1, ENSG00000155657	TTN, ENSG00000232386	RP11-66B24.2, ENSG00000258223	PRSS58, ENSG00000143228	NUF2, ENSG00000053524	MCF2L2, ENSG00000174992	ZG16, ENSG00000235631	RNF148, ENSG00000244300	RP11-475N22.4, ENSG00000178163	ZNF518B, ENSG00000260464	RP4-561L24.3, ENSG00000145908	ZNF300, ENSG00000268107	AC003005.4, ENSG00000258810	RP11-219E7.1, ENSG00000173431	RNASE8, ENSG00000174807	CD248, ENSG00000267259	CTD-2008P7.9, ENSG00000273281	RP11-339B21.14, ENSG00000253644	KB-1930G5.3, ENSG00000272656	RP11-219D15.3, ENSG00000121742	GJB6, ENSG00000253392	AC006277.2, ENSG00000072121	ZFYVE26, ENSG00000238004	AC009227.3, ENSG00000267607	CTD-2369P2.8, ENSG00000255322	RP11-22P4.1, ENSG00000205622	AF064858.6, ENSG00000235192	AC009495.2, ENSG00000151388	ADAMTS12, ENSG00000224739	AC016735.1, ENSG00000242104	RP11-340E6.1, ENSG00000250337	LINC01021, ENSG00000259437	RP11-798K3.4, ENSG00000228526	RP3-510D11.1, ENSG00000155719	OTOA, ENSG00000257787	RP11-20L19.1, ENSG00000160505	NLRP4, ENSG00000114455	HHLA2, ENSG00000235586	AC011247.3, ENSG00000258926	RP11-47I22.1, ENSG00000126785	RHOJ, ENSG00000253524	RP11-89M20.2, ENSG00000272817	RP11-402D21.2, ENSG00000258867	LINC01146, ENSG00000168246	UBTD2, ENSG00000259364	RP11-64K12.9, ENSG00000168903	BTNL3, ENSG00000179941	BBS10, ENSG00000198551	ZNF627, ENSG00000250234	CTD-2024P10.1, ENSG00000232183	RP6-27P15.2, ENSG00000258455	RP11-665C16.5, ENSG00000267552	CTD-2528L19.4, ENSG00000120094	HOXB1, ENSG00000176358	TAC4, ENSG00000273104	CMP21-97G8.1, ENSG00000254902	ANO1-AS1, ENSG00000101435	CST9L, ENSG00000173786	CNP, ENSG00000254187	CTB-78F1.1, ENSG00000248864	RP11-348J24.1, ENSG00000184985	SORCS2, ENSG00000204481	PRAMEF14, ENSG00000235772	AP001625.6, ENSG00000267070	RP11-10K17.6, ENSG00000272040	CTC-366B18.4, ENSG00000158859	ADAMTS4, ENSG00000234551	RP11-123H22.1, ENSG00000231297	SNORD45, ENSG00000118972	FGF23, ENSG00000229272	RP11-498J9.2, ENSG00000177822	AC108142.1, ENSG00000106299	WASL, ENSG00000133115	STOML3, ENSG00000236544	AC008060.8, ENSG00000134871	COL4A2, ENSG00000166869	CHP2, ENSG00000173401	GLIPR1L1, ENSG00000162757	C1orf74, ENSG00
#> PC_ 1 
#> Positive:  ENSG00000063177	RPL18, ENSG00000122026	RPL21, ENSG00000198821	CD247, ENSG00000144746	ARL6IP5, ENSG00000139193	CD27, ENSG00000123240	OPTN, ENSG00000027075	PRKCH, ENSG00000260083	MIR4519, ENSG00000088766	CRLS1, ENSG00000090554	FLT3LG 
#> 	   ENSG00000161204	ABCF3, ENSG00000154473	BUB3, ENSG00000090447	TFAP4, ENSG00000132382	MYBBP1A, ENSG00000155438	NIFK, ENSG00000176248	ANAPC2, ENSG00000105373	GLTSCR2, ENSG00000099917	MED15, ENSG00000016864	GLT8D1, ENSG00000154764	WNT7A 
#> 	   ENSG00000120915	EPHX2, ENSG00000149311	ATM, ENSG00000189339	SLC35E2B, ENSG00000224051	GLTPD1, ENSG00000109920	FNBP4, ENSG00000166971	AKTIP, ENSG00000126749	EMG1, ENSG00000140545	MFGE8, ENSG00000105193	RPS16, ENSG00000124787	RPP40 
#> Negative:  ENSG00000119411	BSPRY, ENSG00000130787	HIP1R, ENSG00000085788	DDHD2, ENSG00000186522	SEPT10, ENSG00000197780	TAF13, ENSG00000156170	NDUFAF6, ENSG00000204388	HSPA1B, ENSG00000135093	USP30, ENSG00000171848	RRM2, ENSG00000164305	CASP3 
#> 	   ENSG00000071655	MBD3, ENSG00000167196	FBXO22, ENSG00000118292	C1orf54, ENSG00000119782	FKBP1B, ENSG00000140691	ARMC5, ENSG00000182359	KBTBD3, ENSG00000100368	CSF2RB, ENSG00000223865	HLA-DPB1, ENSG00000139921	TMX1, ENSG00000075234	TTC38 
#> 	   ENSG00000066336	SPI1, ENSG00000173064	HECTD4, ENSG00000140263	SORD, ENSG00000030582	GRN, ENSG00000213246	SUPT4H1, ENSG00000197816	CCDC180, ENSG00000166145	SPINT1, ENSG00000166224	SGPL1, ENSG00000130714	POMT1, ENSG00000117899	MESDC2 
#> PC_ 2 
#> Positive:  ENSG00000223865	HLA-DPB1, ENSG00000112799	LY86, ENSG00000066336	SPI1, ENSG00000204482	LST1, ENSG00000118292	C1orf54, ENSG00000025708	TYMP, ENSG00000030582	GRN, ENSG00000161243	FBXO27, ENSG00000146757	ZNF92, ENSG00000166736	HTR3A 
#> 	   ENSG00000166145	SPINT1, ENSG00000119782	FKBP1B, ENSG00000092036	HAUS4, ENSG00000139278	GLIPR1, ENSG00000198951	NAGA, ENSG00000124570	SERPINB6, ENSG00000101220	C20orf27, ENSG00000163131	CTSS, ENSG00000140263	SORD, ENSG00000126759	CFP 
#> 	   ENSG00000124357	NAGK, ENSG00000115204	MPV17, ENSG00000139921	TMX1, ENSG00000155366	RHOC, ENSG00000130475	FCHO1, ENSG00000100908	EMC9, ENSG00000155330	C16orf87, ENSG00000168303	MPLKIP, ENSG00000128524	ATP6V1F, ENSG00000169981	ZNF35 
#> Negative:  ENSG00000204388	HSPA1B, ENSG00000197780	TAF13, ENSG00000085788	DDHD2, ENSG00000135093	USP30, ENSG00000119411	BSPRY, ENSG00000071655	MBD3, ENSG00000156170	NDUFAF6, ENSG00000171848	RRM2, ENSG00000167196	FBXO22, ENSG00000186522	SEPT10 
#> 	   ENSG00000130787	HIP1R, ENSG00000164305	CASP3, ENSG00000160113	NR2F6, ENSG00000184451	CCR10, ENSG00000150347	ARID5B, ENSG00000197816	CCDC180, ENSG00000100368	CSF2RB, ENSG00000075234	TTC38, ENSG00000055163	CYFIP2, ENSG00000128829	EIF2AK4 
#> 	   ENSG00000134285	FKBP11, ENSG00000139193	CD27, ENSG00000103353	UBFD1, ENSG00000108349	CASC3, ENSG00000004468	CD38, ENSG00000090554	FLT3LG, ENSG00000180096	SEPT1, ENSG00000100124	ANKRD54, ENSG00000130714	POMT1, ENSG00000115806	GORASP2 
#> PC_ 3 
#> Positive:  ENSG00000229084	LINC00226, ENSG00000263508	RP11-963H4.3, ENSG00000183018	SPNS2, ENSG00000177398	UMODL1, ENSG00000107447	DNTT, ENSG00000175893	ZDHHC21, ENSG00000174059	CD34, ENSG00000269996	RP11-343N15.5, ENSG00000167483	FAM129C, ENSG00000048991	R3HDM1 
#> 	   ENSG00000214367	HAUS3, ENSG00000032742	IFT88, ENSG00000159753	RLTPR, ENSG00000184708	EIF4ENIF1, ENSG00000185291	IL3RA, ENSG00000121931	LRIF1, ENSG00000138138	ATAD1, ENSG00000256060	TRAPPC2P1, ENSG00000215039	CD27-AS1, ENSG00000082641	NFE2L1 
#> 	   ENSG00000145687	SSBP2, ENSG00000106608	URGCP, ENSG00000115289	PCGF1, ENSG00000155252	PI4K2A, ENSG00000143570	SLC39A1, ENSG00000117335	CD46, ENSG00000251022	THAP9-AS1, ENSG00000271858	XXcos-LUCA11.4, ENSG00000148248	SURF4, ENSG00000148719	DNAJB12 
#> Negative:  ENSG00000025708	TYMP, ENSG00000229117	RPL41, ENSG00000163131	CTSS, ENSG00000030582	GRN, ENSG00000112799	LY86, ENSG00000134072	CAMK1, ENSG00000120742	SERP1, ENSG00000164733	CTSB, ENSG00000026025	VIM, ENSG00000174886	NDUFA11 
#> 	   ENSG00000186010	NDUFA13, ENSG00000182578	CSF1R, ENSG00000172216	CEBPB, ENSG00000173221	GLRX, ENSG00000075234	TTC38, ENSG00000223865	HLA-DPB1, ENSG00000124357	NAGK, ENSG00000108826	MRPL27, ENSG00000119705	SLIRP, ENSG00000108349	CASC3 
#> 	   ENSG00000172757	CFL1, ENSG00000204482	LST1, ENSG00000177663	IL17RA, ENSG00000185507	IRF7, ENSG00000136152	COG3, ENSG00000156170	NDUFAF6, ENSG00000204388	HSPA1B, ENSG00000197780	TAF13, ENSG00000085788	DDHD2, ENSG00000119411	BSPRY 
#> PC_ 4 
#> Positive:  ENSG00000136717	BIN1, ENSG00000100129	EIF3L, ENSG00000146757	ZNF92, ENSG00000161243	FBXO27, ENSG00000166736	HTR3A, ENSG00000139193	CD27, ENSG00000140263	SORD, ENSG00000105373	GLTSCR2, ENSG00000119782	FKBP1B, ENSG00000133250	ZNF414 
#> 	   ENSG00000118292	C1orf54, ENSG00000115042	FAHD2A, ENSG00000111348	ARHGDIB, ENSG00000198821	CD247, ENSG00000122026	RPL21, ENSG00000183431	SF3A3, ENSG00000164327	RICTOR, ENSG00000145741	BTF3, ENSG00000090554	FLT3LG, ENSG00000162244	RPL29 
#> 	   ENSG00000164088	PPM1M, ENSG00000139921	TMX1, ENSG00000171148	TADA3, ENSG00000136819	C9orf78, ENSG00000015479	MATR3, ENSG00000184986	TMEM121, ENSG00000168283	BMI1, ENSG00000169981	ZNF35, ENSG00000114923	SLC4A3, ENSG00000123908	AGO2 
#> Negative:  ENSG00000204482	LST1, ENSG00000025708	TYMP, ENSG00000126759	CFP, ENSG00000223865	HLA-DPB1, ENSG00000163131	CTSS, ENSG00000066336	SPI1, ENSG00000134072	CAMK1, ENSG00000167851	CD300A, ENSG00000136867	SLC31A2, ENSG00000128524	ATP6V1F 
#> 	   ENSG00000030582	GRN, ENSG00000184708	EIF4ENIF1, ENSG00000159753	RLTPR, ENSG00000172216	CEBPB, ENSG00000121064	SCPEP1, ENSG00000102524	TNFSF13B, ENSG00000272831	RP11-792A8.4, ENSG00000119705	SLIRP, ENSG00000115233	PSMD14, ENSG00000182578	CSF1R 
#> 	   ENSG00000158481	CD1C, ENSG00000173442	EHBP1L1, ENSG00000229117	RPL41, ENSG00000196642	RABL6, ENSG00000249222	ATP5L2, ENSG00000106003	LFNG, ENSG00000124357	NAGK, ENSG00000099956	SMARCB1, ENSG00000267519	MIR24-2, ENSG00000136152	COG3 
#> PC_ 5 
#> Positive:  ENSG00000161243	FBXO27, ENSG00000166736	HTR3A, ENSG00000146757	ZNF92, ENSG00000114923	SLC4A3, ENSG00000169981	ZNF35, ENSG00000135392	DNAJC14, ENSG00000123908	AGO2, ENSG00000168283	BMI1, ENSG00000166145	SPINT1, ENSG00000164327	RICTOR 
#> 	   ENSG00000092036	HAUS4, ENSG00000170921	TANC2, ENSG00000129521	EGLN3, ENSG00000247240	UBL7-AS1, ENSG00000087076	HSD17B14, ENSG00000238227	C9orf69, ENSG00000128039	SRD5A3, ENSG00000143797	MBOAT2, ENSG00000100983	GSS, ENSG00000126821	SGPP1 
#> 	   ENSG00000131368	MRPS25, ENSG00000155330	C16orf87, ENSG00000120868	APAF1, ENSG00000261737	RP4-612B15.3, ENSG00000184575	XPOT, ENSG00000173559	NABP1, ENSG00000110921	MVK, ENSG00000110367	DDX6, ENSG00000104164	BLOC1S6, ENSG00000214756	METTL12 
#> Negative:  ENSG00000173064	HECTD4, ENSG00000080200	CRYBG3, ENSG00000170542	SERPINB9, ENSG00000106683	LIMK1, ENSG00000272269	RP11-500C11.3, ENSG00000172831	CES2, ENSG00000163040	CCDC74A, ENSG00000100416	TRMU, ENSG00000167536	DHRS13, ENSG00000236104	ZBTB22 
#> 	   ENSG00000150776	C11orf57, ENSG00000004487	KDM1A, ENSG00000198498	TMA16, ENSG00000156876	SASS6, ENSG00000140090	SLC24A4, ENSG00000119782	FKBP1B, ENSG00000146067	FAM193B, ENSG00000254415	SIGLEC14, ENSG00000167522	ANKRD11, ENSG00000125817	CENPB 
#> 	   ENSG00000182359	KBTBD3, ENSG00000121644	DESI2, ENSG00000133250	ZNF414, ENSG00000136689	IL1RN, ENSG00000184674	GSTT1, ENSG00000137074	APTX, ENSG00000148719	DNAJB12, ENSG00000140691	ARMC5, ENSG00000131759	RARA, ENSG00000198715	C1orf85 
#> Computing nearest neighbor graph
#> Computing SNN
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 136
#> Number of edges: 4813
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.7201
#> Number of communities: 2
#> Elapsed time: 0 seconds
#> Normalizing layer: counts
#> Finding variable features for layer counts
#> Warning: pseudoinverse used at -2.5336
#> Warning: neighborhood radius 0.49481
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.090619
#> Centering and scaling data matrix
#> Warning: The following 971 features requested have zero variance; running reduction without them: ENSG00000172733	PURG, ENSG00000174529	TMEM81, ENSG00000261231	RP11-523L20.2, ENSG00000189367	KIAA0408, ENSG00000086570	FAT2, ENSG00000249882	RP11-665I14.1, ENSG00000237853	RP5-833A20.1, ENSG00000255276	RRM1-AS1, ENSG00000224902	GAGE12H, ENSG00000164185	ZNF474, ENSG00000248703	RP11-495K9.3, ENSG00000251151	HOXC-AS3, ENSG00000177076	ACER2, ENSG00000267038	RP11-1096D5.2, ENSG00000228569	AC073133.2, ENSG00000260780	RP11-580I1.1, ENSG00000231703	RP4-669H2.1, ENSG00000224658	RP11-631F7.1, ENSG00000250431	RP13-577H12.2, ENSG00000169398	PTK2, ENSG00000231731	AC010976.2, ENSG00000255420	RP11-236J17.3, ENSG00000204889	KRT40, ENSG00000237907	RP5-1013A22.2, ENSG00000185792	NLRP9, ENSG00000272579	RP11-101E13.5, ENSG00000272121	RP4-555D20.4, ENSG00000196754	S100A2, ENSG00000169679	BUB1, ENSG00000251532	CTD-2245E15.3, ENSG00000125954	CHURC1-FNTB, ENSG00000168621	GDNF, ENSG00000224844	AC107079.1, ENSG00000139988	RDH12, ENSG00000132693	CRP, ENSG00000174032	SLC25A30, ENSG00000257985	RP1-228P16.4, ENSG00000261583	CTD-2385L22.1, ENSG00000172260	NEGR1, ENSG00000230687	RP5-1114G22.2, ENSG00000261789	RP11-709D24.5, ENSG00000065320	NTN1, ENSG00000267467	APOC4, ENSG00000006788	MYH13, ENSG00000234622	RP13-297E16.5, ENSG00000233234	RP11-740P5.2, ENSG00000269064	AL160286.1, ENSG00000269895	AP000654.4, ENSG00000267580	CTD-2540B15.11, ENSG00000265778	RP11-17M16.2, ENSG00000269199	AC003973.5, ENSG00000184492	FOXD4L1, ENSG00000249847	RP11-192C21.2, ENSG00000226113	RP11-672F9.1, ENSG00000255104	ZNF286A, ENSG00000263508	RP11-963H4.3, ENSG00000231294	AC006994.2, ENSG00000064309	CDON, ENSG00000254290	RP11-150O12.3, ENSG00000261717	RP11-77K12.1, ENSG00000184454	NCMAP, ENSG00000228539	RP11-526K17.2, ENSG00000255062	RP11-712L6.5, ENSG00000236306	RP11-33G16.1, ENSG00000204187	LINC00619, ENSG00000120436	GPR31, ENSG00000227131	RP11-552E20.1, ENSG00000245385	RP11-334E6.10, ENSG00000226386	PARD3-AS1, ENSG00000165309	ARMC3, ENSG00000170835	CEL, ENSG00000256391	SDIM1, ENSG00000232530	RP1-102K2.6, ENSG00000185894	BPY2C, ENSG00000198400	NTRK1, ENSG00000258888	RP11-326A13.1, ENSG00000235085	AC091153.4, ENSG00000204702	OR2J1, ENSG00000163251	FZD5, ENSG00000203301	AL590822.1, ENSG00000178412	RP11-567M16.3, ENSG00000233835	RP1-92C4.2, ENSG00000268012	AC016559.1, ENSG00000204252	HLA-DOA, ENSG00000268549	RP11-295P9.8, ENSG00000175121	WFDC5, ENSG00000259063	DUX4L5, ENSG00000228213	NLGN1-AS1, ENSG00000233821	ENOX1-AS1, ENSG00000139514	SLC7A1, ENSG00000221938	OR2A14, ENSG00000091128	LAMB4, ENSG00000224691	GS1-174L6.4, ENSG00000183032	SLC25A21, ENSG00000259312	RP11-522B15.5, ENSG00000263335	AF001548.5, ENSG00000145198	VWA5B2, ENSG00000174111	SOCS7, ENSG00000100012	SEC14L3, ENSG00000181074	OR52N4, ENSG00000250241	RP11-9G1.3, ENSG00000236761	CTAGE9, ENSG00000103528	SYT17, ENSG00000248828	RP11-673E1.4, ENSG00000204110	RP1-153P14.8, ENSG00000227518	MIR1302-9, ENSG00000224201	PNMA6A, ENSG00000225216	AC007362.3, ENSG00000251258	RFPL4B, ENSG00000249639	CTB-138E5.1, ENSG00000156486	KCNS2, ENSG00000228272	AC106874.1, ENSG00000163914	RHO, ENSG00000196532	HIST1H3C, ENSG00000256802	RP11-680F8.1, ENSG00000134853	PDGFRA, ENSG00000228214	LINC00693, ENSG00000205537	RP11-89H19.1, ENSG00000229855	CTC-546K23.1, ENSG00000236463	LINC00427, ENSG00000238058	RP11-432J22.2, ENSG00000259044	RP11-789A21.1, ENSG00000268091	AC078925.1, ENSG00000254768	CTD-2140G10.2, ENSG00000125257	ABCC4, ENSG00000268126	AC091948.1, ENSG00000272762	RP11-155D18.14, ENSG00000164236	ANKRD33B, ENSG00000165066	NKX6-3, ENSG00000172382	PRSS27, ENSG00000232529	RP1-257I9.2, ENSG00000222030	AC007131.1, ENSG00000203364	RP11-370F5.4, ENSG00000240553	RP1-184J9.2, ENSG00000099812	MISP, ENSG00000253893	FAM85B, ENSG00000176742	OR51V1, ENSG00000259518	RP11-597K23.2, ENSG00000069764	PLA2G10, ENSG00000253735	RP11-115J16.3, ENSG00000230920	RP11-527D15.1, ENSG00000262576	PCDHGA4, ENSG00000130433	CACNG6, ENSG00000004776	HSPB6, ENSG00000205693	MANSC4, ENSG00000273091	AP000255.6, ENSG00000143797	MBOAT2, ENSG00000170549	IRX1, ENSG00000261272	MUC22, ENSG00000227782	AC002553.1, ENSG00000115602	IL1RL1, ENSG00000176754	LINC00303, ENSG00000131480	AOC2, ENSG00000140798	ABCC12, ENSG00000164855	TMEM184A, ENSG00000233670	PIRT, ENSG00000007171	NOS2, ENSG00000226957	RP4-533D7.4, ENSG00000187730	GABRD, ENSG00000259375	RP11-815J21.2, ENSG00000255327	RP11-555G19.1, ENSG00000099377	HSD3B7, ENSG00000267769	CTB-50L17.9, ENSG00000227047	AC012065.4, ENSG00000206043	C18orf63, ENSG00000165985	C1QL3, ENSG00000231394	RP11-310H4.3, ENSG00000226209	AL138764.1, ENSG00000267191	RP11-15A1.2, ENSG00000198812	LRRC10, ENSG00000255313	RP11-1082L8.2, ENSG00000268809	AC012360.2, ENSG00000231426	RP5-899B16.1, ENSG00000224616	RP11-305E17.6, ENSG00000259714	LINC00594, ENSG00000254084	KB-1930G5.4, ENSG00000246331	RP11-77I22.2, ENSG00000130701	RBBP8NL, ENSG00000269676	MUC8, ENSG00000268472	AP002884.2, ENSG00000268923	AL645922.1, ENSG00000139797	RNF113B, ENSG00000171540	OTP, ENSG00000256209	RP11-897M7.1, ENSG00000257894	RP1-78O14.1, ENSG00000182578	CSF1R, ENSG00000264750	RP11-277J6.2, ENSG00000130822	PNCK, ENSG00000058085	LAMC2, ENSG00000167858	TEKT1, ENSG00000260235	CTD-2026K11.3, ENSG00000271973	RP11-572O6.1, ENSG00000232080	XXbac-BPG254F23.7, ENSG00000231171	LINC01098, ENSG00000226310	RP3-323P24.3, ENSG00000225872	AC002398.5, ENSG00000260201	RP11-143N13.2, ENSG00000260725	AC005307.1, ENSG00000155657	TTN, ENSG00000232386	RP11-66B24.2, ENSG00000258223	PRSS58, ENSG00000174992	ZG16, ENSG00000235631	RNF148, ENSG00000244300	RP11-475N22.4, ENSG00000178163	ZNF518B, ENSG00000268107	AC003005.4, ENSG00000258810	RP11-219E7.1, ENSG00000173431	RNASE8, ENSG00000174807	CD248, ENSG00000267259	CTD-2008P7.9, ENSG00000273281	RP11-339B21.14, ENSG00000229084	LINC00226, ENSG00000253644	KB-1930G5.3, ENSG00000254415	SIGLEC14, ENSG00000272656	RP11-219D15.3, ENSG00000121742	GJB6, ENSG00000253392	AC006277.2, ENSG00000238004	AC009227.3, ENSG00000267607	CTD-2369P2.8, ENSG00000255322	RP11-22P4.1, ENSG00000205622	AF064858.6, ENSG00000151388	ADAMTS12, ENSG00000224739	AC016735.1, ENSG00000242104	RP11-340E6.1, ENSG00000250337	LINC01021, ENSG00000259437	RP11-798K3.4, ENSG00000228526	RP3-510D11.1, ENSG00000162711	NLRP3, ENSG00000155719	OTOA, ENSG00000257787	RP11-20L19.1, ENSG00000160505	NLRP4, ENSG00000114455	HHLA2, ENSG00000235586	AC011247.3, ENSG00000258926	RP11-47I22.1, ENSG00000126785	RHOJ, ENSG00000159556	ISL2, ENSG00000253524	RP11-89M20.2, ENSG00000225265	RP11-378J18.3, ENSG00000273314	RP5-1136G13.2, ENSG00000258867	LINC01146, ENSG00000168246	UBTD2, ENSG00000259364	RP11-64K12.9, ENSG00000168903	BTNL3, ENSG00000179941	BBS10, ENSG00000250234	CTD-2024P10.1, ENSG00000232183	RP6-27P15.2, ENSG00000258455	RP11-665C16.5, ENSG00000267552	CTD-2528L19.4, ENSG00000120094	HOXB1, ENSG00000176358	TAC4, ENSG00000273104	CMP21-97G8.1, ENSG00000254902	ANO1-AS1, ENSG00000101435	CST9L, ENSG00000254187	CTB-78F1.1, ENSG00000248864	RP11-348J24.1, ENSG00000184985	SORCS2, ENSG00000204481	PRAMEF14, ENSG00000235772	AP001625.6, ENSG00000267070	RP11-10K17.6, ENSG00000272040	CTC-366B18.4, ENSG00000158859	ADAMTS4, ENSG00000234551	RP11-123H22.1, ENSG00000231297	SNORD45, ENSG00000118972	FGF23, ENSG00000229272	RP11-498J9.2, ENSG00000177822	AC108142.1, ENSG00000106299	WASL, ENSG00000133115	STOML3, ENSG00000236544	AC008060.8, ENSG00000134871	COL4A2, ENSG00000166869	CHP2, ENSG00000173401	GLIPR1L1, ENSG00000162757	C1orf74, ENSG00000261194	RP11-704C2.1, ENSG00000255355	AP000640.2, ENSG00000267838	AC008746.12, ENSG00000215549	AL138815.1, ENSG00000187612	OR5W2, ENSG00000238042	RP11-815M8.1, ENSG00000269709	AL109659.1, ENSG00000177476	OR2G3, ENSG00000188859	FAM78B, ENSG00000228302	RP11-186N15.3, ENSG00000272962	SLC5A3, ENSG00000257523	CTD-2384A14.2, ENSG00000237189	RP11-85G21.2, ENSG00000261020	RP11-744K17.1, ENSG00000226516	FAM138B, ENSG00000260977	RP11-333I13.1, ENSG00000241732	RP11-38P22.2, ENSG00000255618	RP11-357K6.1, ENSG00000134588	USP26, ENSG00000234193	AC097461.4, ENSG00000245870	LINC00682, ENSG00000268842	CTD-2245F17.2, ENSG00000267601	RP11-323N12.5, ENSG00000215635	AC145205.1, ENSG00000240710	RP11-430C7.4, ENSG00000234279	AC113618.1, ENSG00000204001	LCN8, ENSG00000260905	
#> PC_ 1 
#> Positive:  ENSG00000187118	CMC1, ENSG00000198821	CD247, ENSG00000155366	RHOC, ENSG00000156411	C14orf2, ENSG00000168701	TMEM208, ENSG00000075234	TTC38, ENSG00000262202	SNORD3D, ENSG00000167851	CD300A, ENSG00000184007	PTP4A2, ENSG00000125743	SNRPD2 
#> 	   ENSG00000186010	NDUFA13, ENSG00000213741	RPS29, ENSG00000099956	SMARCB1, ENSG00000130770	ATPIF1, ENSG00000196262	PPIA, ENSG00000101474	APMAP, ENSG00000119705	SLIRP, ENSG00000141002	TCF25, ENSG00000088986	DYNLL1, ENSG00000198574	SH2D1B 
#> 	   ENSG00000170540	ARL6IP1, ENSG00000105058	FAM32A, ENSG00000115053	NCL, ENSG00000108826	MRPL27, ENSG00000139725	RHOF, ENSG00000004468	CD38, ENSG00000122482	ZNF644, ENSG00000184752	NDUFA12, ENSG00000136527	TRA2B, ENSG00000120699	EXOSC8 
#> Negative:  ENSG00000122026	RPL21, ENSG00000139193	CD27, ENSG00000090554	FLT3LG, ENSG00000120915	EPHX2, ENSG00000162244	RPL29, ENSG00000026025	VIM, ENSG00000135916	ITM2C, ENSG00000100129	EIF3L, ENSG00000105373	GLTSCR2, ENSG00000063177	RPL18 
#> 	   ENSG00000139197	PEX5, ENSG00000105193	RPS16, ENSG00000167642	SPINT2, ENSG00000227039	ITGB2-AS1, ENSG00000197880	MDS2, ENSG00000128271	ADORA2A, ENSG00000134202	GSTM3, ENSG00000266947	RP11-799D4.4, ENSG00000159496	RGL4, ENSG00000153391	INO80C 
#> 	   ENSG00000232442	CTD-3184A7.4, ENSG00000105443	CYTH2, ENSG00000173327	MAP3K11, ENSG00000154764	WNT7A, ENSG00000121749	TBC1D15, ENSG00000102967	DHODH, ENSG00000095059	DHPS, ENSG00000168209	DDIT4, ENSG00000198355	PIM3, ENSG00000126759	CFP 
#> PC_ 2 
#> Positive:  ENSG00000167136	ENDOG, ENSG00000148218	ALAD, ENSG00000027075	PRKCH, ENSG00000180209	MYLPF, ENSG00000173457	PPP1R14B, ENSG00000109534	GAR1, ENSG00000110367	DDX6, ENSG00000198799	LRIG2, ENSG00000167232	ZNF91, ENSG00000150347	ARID5B 
#> 	   ENSG00000121644	DESI2, ENSG00000105829	BET1, ENSG00000224051	GLTPD1, ENSG00000130770	ATPIF1, ENSG00000158828	PINK1, ENSG00000113595	TRIM23, ENSG00000136819	C9orf78, ENSG00000185305	ARL15, ENSG00000213741	RPS29, ENSG00000156411	C14orf2 
#> 	   ENSG00000004468	CD38, ENSG00000175756	AURKAIP1, ENSG00000105676	ARMC6, ENSG00000152229	PSTPIP2, ENSG00000185760	KCNQ5, ENSG00000198574	SH2D1B, ENSG00000114416	FXR1, ENSG00000163564	PYHIN1, ENSG00000140987	ZSCAN32, ENSG00000262202	SNORD3D 
#> Negative:  ENSG00000112640	PPP2R5D, ENSG00000163577	EIF5A2, ENSG00000224597	PTCHD3P1, ENSG00000156853	ZNF689, ENSG00000164305	CASP3, ENSG00000064651	SLC12A2, ENSG00000113391	FAM172A, ENSG00000166260	COX11, ENSG00000169499	PLEKHA2, ENSG00000153767	GTF2E1 
#> 	   ENSG00000136152	COG3, ENSG00000134285	FKBP11, ENSG00000153107	ANAPC1, ENSG00000130787	HIP1R, ENSG00000259230	CTD-2555C10.3, ENSG00000185507	IRF7, ENSG00000099942	CRKL, ENSG00000272269	RP11-500C11.3, ENSG00000231889	TRAF3IP2-AS1, ENSG00000102900	NUP93 
#> 	   ENSG00000026025	VIM, ENSG00000160781	PAQR6, ENSG00000152193	RNF219, ENSG00000241370	RPP21, ENSG00000124787	RPP40, ENSG00000125637	PSD4, ENSG00000170100	ZNF778, ENSG00000254470	AP5B1, ENSG00000142751	GPN2, ENSG00000156313	RPGR 
#> PC_ 3 
#> Positive:  ENSG00000166228	PCBD1, ENSG00000141522	ARHGDIA, ENSG00000109534	GAR1, ENSG00000123472	ATPAF1, ENSG00000121644	DESI2, ENSG00000013288	MAN2B2, ENSG00000184731	FAM110C, ENSG00000026025	VIM, ENSG00000272269	RP11-500C11.3, ENSG00000110048	OSBP 
#> 	   ENSG00000131759	RARA, ENSG00000111666	CHPT1, ENSG00000204856	FAM216A, ENSG00000099991	CABIN1, ENSG00000197223	C1D, ENSG00000174100	MRPL45, ENSG00000165996	PTPLA, ENSG00000136147	PHF11, ENSG00000164327	RICTOR, ENSG00000169905	TOR1AIP2 
#> 	   ENSG00000143314	MRPL24, ENSG00000141101	NOB1, ENSG00000120697	ALG5, ENSG00000198498	TMA16, ENSG00000106367	AP1S1, ENSG00000171148	TADA3, ENSG00000148840	PPRC1, ENSG00000253797	UTP14C, ENSG00000167483	FAM129C, ENSG00000122481	RWDD3 
#> Negative:  ENSG00000139278	GLIPR1, ENSG00000198521	ZNF43, ENSG00000166326	TRIM44, ENSG00000125482	TTF1, ENSG00000224051	GLTPD1, ENSG00000049167	ERCC8, ENSG00000237441	RGL2, ENSG00000113595	TRIM23, ENSG00000143815	LBR, ENSG00000138413	IDH1 
#> 	   ENSG00000068745	IP6K2, ENSG00000168818	STX18, ENSG00000004468	CD38, ENSG00000099849	RASSF7, ENSG00000145414	NAF1, ENSG00000075089	ACTR6, ENSG00000204564	C6orf136, ENSG00000153487	ING1, ENSG00000147536	GINS4, ENSG00000164305	CASP3 
#> 	   ENSG00000163564	PYHIN1, ENSG00000162959	MEMO1, ENSG00000116906	GNPAT, ENSG00000186017	ZNF566, ENSG00000152229	PSTPIP2, ENSG00000032742	IFT88, ENSG00000260083	MIR4519, ENSG00000134516	DOCK2, ENSG00000153814	JAZF1, ENSG00000180957	PITPNB 
#> PC_ 4 
#> Positive:  ENSG00000172216	CEBPB, ENSG00000105705	SUGP1, ENSG00000136811	ODF2, ENSG00000150776	C11orf57, ENSG00000071655	MBD3, ENSG00000113595	TRIM23, ENSG00000148218	ALAD, ENSG00000130787	HIP1R, ENSG00000013288	MAN2B2, ENSG00000129534	MIS18BP1 
#> 	   ENSG00000228857	AC104653.1, ENSG00000223759	RP11-353N4.4, ENSG00000174456	C12orf76, ENSG00000120699	EXOSC8, ENSG00000174100	MRPL45, ENSG00000153107	ANAPC1, ENSG00000164305	CASP3, ENSG00000167862	ICT1, ENSG00000172831	CES2, ENSG00000184277	TM2D3 
#> 	   ENSG00000165359	DDX26B, ENSG00000032742	IFT88, ENSG00000160285	LSS, ENSG00000164306	PRIMPOL, ENSG00000197841	ZNF181, ENSG00000260083	MIR4519, ENSG00000122026	RPL21, ENSG00000116497	S100PBP, ENSG00000224051	GLTPD1, ENSG00000135521	LTV1 
#> Negative:  ENSG00000273344	PAXIP1-AS1, ENSG00000143167	GPA33, ENSG00000102900	NUP93, ENSG00000108010	GLRX3, ENSG00000128829	EIF2AK4, ENSG00000272269	RP11-500C11.3, ENSG00000092201	SUPT16H, ENSG00000160781	PAQR6, ENSG00000254470	AP5B1, ENSG00000126756	UXT 
#> 	   ENSG00000108774	RAB5C, ENSG00000178149	DALRD3, ENSG00000127616	SMARCA4, ENSG00000099942	CRKL, ENSG00000148335	NTMT1, ENSG00000140545	MFGE8, ENSG00000134516	DOCK2, ENSG00000180423	HARBI1, ENSG00000075234	TTC38, ENSG00000267680	ZNF224 
#> 	   ENSG00000173933	RBM4, ENSG00000180096	SEPT1, ENSG00000114416	FXR1, ENSG00000132549	VPS13B, ENSG00000168090	COPS6, ENSG00000215014	AL645728.1, ENSG00000262202	SNORD3D, ENSG00000166548	TK2, ENSG00000142856	ITGB3BP, ENSG00000105058	FAM32A 
#> PC_ 5 
#> Positive:  ENSG00000180376	CCDC66, ENSG00000121064	SCPEP1, ENSG00000137404	NRM, ENSG00000165671	NSD1, ENSG00000166133	RPUSD2, ENSG00000141002	TCF25, ENSG00000120948	TARDBP, ENSG00000174886	NDUFA11, ENSG00000105373	GLTSCR2, ENSG00000126749	EMG1 
#> 	   ENSG00000221916	C19orf73, ENSG00000122971	ACADS, ENSG00000126457	PRMT1, ENSG00000144824	PHLDB2, ENSG00000140545	MFGE8, ENSG00000135521	LTV1, ENSG00000107771	CCSER2, ENSG00000147874	HAUS6, ENSG00000164733	CTSB, ENSG00000103671	TRIP4 
#> 	   ENSG00000136717	BIN1, ENSG00000273080	RP11-301O19.1, ENSG00000168701	TMEM208, ENSG00000108826	MRPL27, ENSG00000067955	CBFB, ENSG00000183647	ZNF530, ENSG00000117000	RLF, ENSG00000165983	PTER, ENSG00000104047	DTWD1, ENSG00000167232	ZNF91 
#> Negative:  ENSG00000153107	ANAPC1, ENSG00000153140	CETN3, ENSG00000181035	SLC25A42, ENSG00000164306	PRIMPOL, ENSG00000064651	SLC12A2, ENSG00000147536	GINS4, ENSG00000163577	EIF5A2, ENSG00000197568	HHLA3, ENSG00000197157	SND1, ENSG00000241839	PLEKHO2 
#> 	   ENSG00000156170	NDUFAF6, ENSG00000169499	PLEKHA2, ENSG00000224597	PTCHD3P1, ENSG00000153767	GTF2E1, ENSG00000187514	PTMA, ENSG00000136169	SETDB2, ENSG00000099330	OCEL1, ENSG00000215039	CD27-AS1, ENSG00000149311	ATM, ENSG00000273117	AC144652.1 
#> 	   ENSG00000198677	TTC37, ENSG00000140307	GTF2A2, ENSG00000130787	HIP1R, ENSG00000125995	ROMO1, ENSG00000148248	SURF4, ENSG00000223759	RP11-353N4.4, ENSG00000135365	PHF21A, ENSG00000158805	ZNF276, ENSG00000259230	CTD-2555C10.3, ENSG00000145414	NAF1 
#> Computing nearest neighbor graph
#> Computing SNN
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 328
#> Number of edges: 22343
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.5141
#> Number of communities: 2
#> Elapsed time: 0 seconds