Find the nearest neighbors

Find the nearest neighbors

FindNNDist(fullcluster, normCount, meaningn = 20)

Arguments

fullcluster

A list of clusters that generated by the function GetCluster().

normCount

A list of normalized gene count matrix generated by the function NormData().

meaningn

default to be 20

Value

A list of distance vectors

Examples

data(sim_result)
# Create example data for fullcluster (mock data)
# fullcluster <- GetCluster(seuratlist)
# Create example data for normCount (mock data)
# normCount <- NormData(seuratlist)
# Define meaningn
meaningn <- 20

FindNNDist(sim_result[[1]], sim_result[[2]], meaningn = meaningn)
#>   |                                                                    |                                                            |   0%  |                                                                    |====================                                        |  33%  |                                                                    |========================================                    |  67%  |                                                                    |============================================================| 100%
#> $`sample-1`
#> $`sample-1`[[1]]
#> $`sample-1`[[1]]$`Cluster-1-dist1`
#> $`sample-1`[[1]]$`Cluster-1-dist1`[[1]]
#> ATTTGCACCTATGG-1 ATTAACGACCTAAG-1 TCAATAGAAGTAGA-1 CGAGCCGATTCGTT-1 
#>        0.3820215        0.3824480        0.3996782        0.4037947 
#> CTGCAGCTGAGATA-1 CTCGCATGAGAATG-1 TACGTTACGGACGA-1 CTAGATCTGCAGTT-1 
#>        0.4215034        0.4228631        0.4233883        0.4279768 
#> TGTGAGTGCTTCGC-1 TCATGTACTGCTGA-1 ACTATCACCACTAG-1 GCCAACCTTTTCAC-1 
#>        0.4311639        0.4348488        0.4374487        0.4376245 
#> TTAGGTCTGCTGAT-1 AAAGACGACCAATG-1 GAAGCTTGATTCTC-1 AGTTTCACAGAATG-1 
#>        0.4388135        0.4430774        0.4454401        0.4541111 
#> AGTGAAGAGAGAGC-1 TATACAGATTGACG-1 AGCGAACTGAGGTG-1 ACGGAGGATAGACC-1 
#>        0.4563066        0.4567082        0.4568603        0.4648187 
#> 
#> 
#> $`sample-1`[[1]]$`Cluster-2-dist1`
#> $`sample-1`[[1]]$`Cluster-2-dist1`[[1]]
#> AGCGGCACGGAGTG-1 CTGAGAACGGGACA-1 AATGAGGAGGTAGG-1 ACCTGAGACCTATT-1 
#>        0.5570579        0.5686847        0.5711682        0.5750063 
#> CGATCCACTGGCAT-1 TAGATCCTAGTACC-1 AGACTGACTTTCGT-1 CTGCAGCTTCTTAC-1 
#>        0.5781347        0.5798186        0.5862883        0.5871305 
#> TATACAGATCCTAT-1 GGACATTGGGAGGT-1 GCAGCTCTAGCATC-1 TCTAACTGCTGAAC-1 
#>        0.5884774        0.5894309        0.5903974        0.5906247 
#> CATTGTTGCGGGAA-1 AGCGGCTGCCGTTC-1 ATTAGTGAGGAGCA-1 CAATATGATGGTGT-1 
#>        0.5910974        0.5949419        0.5952026        0.5969553 
#> ATTAACGATCCAAG-1 TATGTGCTAACGAA-1 GGATGTTGAGCACT-1 GTACAGTGTGGTAC-1 
#>        0.5998850        0.6000245        0.6030984        0.6036150 
#> 
#> 
#> $`sample-1`[[1]]$`Cluster-3-dist1`
#> $`sample-1`[[1]]$`Cluster-3-dist1`[[1]]
#> CCAATGGAAAAGCA-1 CCAAAGTGGTACGT-1 CGGACCGACTCCAC-1 TAAAGACTGGGAGT-1 
#>        0.5285320        0.5537157        0.5560560        0.5586981 
#> GAAGAATGGATGAA-1 ACTATCACCTCTAT-1 ATATAGTGTAAGCC-1 TGCAAGACCTCCCA-1 
#>        0.5618884        0.5633627        0.5648629        0.5669600 
#> ACGACAACGCATCA-1 ATTGAAACATCACG-1 GGTGATACGATAAG-1 TTCTACGACTGTAG-1 
#>        0.5703625        0.5758484        0.5758860        0.5813088 
#> GAAGATGATACAGC-1 ACCGTGCTACCTGA-1 ATTGGGTGGCTCCT-1 CAATATGATGGTGT-1 
#>        0.5820775        0.5859982        0.5875378        0.5876409 
#> TAATCCACCTCAGA-1 AACAGCACCTCGCT-1 CGGATAACCACAAC-1 CCAGGTCTTGGGAG-1 
#>        0.5882652        0.5896222        0.5932504        0.5944773 
#> 
#> 
#> $`sample-1`[[1]]$`Cluster-4-dist1`
#> $`sample-1`[[1]]$`Cluster-4-dist1`[[1]]
#> CGATACGAGTAGGG-1 CACCGTACCTGATG-1 ACTCCCGATACTCT-1 ACCTGAGACCTATT-1 
#>        0.4397086        0.4539032        0.4569474        0.4698320 
#> GTCACAGAGAGACG-1 TAGGTCGAACCCTC-1 TACGCAGAGCCCTT-1 TACTCTGAGTCTTT-1 
#>        0.4753724        0.4880795        0.4913024        0.4916527 
#> CAATATGATGGTGT-1 ACTTGACTCACCAA-1 CGATCCACTGGCAT-1 TCGCACTGCTGACA-1 
#>        0.4918639        0.4980750        0.4985978        0.5032653 
#> GCAGCTCTAGCATC-1 GCCAACCTTTTCAC-1 GATCGTGACATCAG-1 TCTAACTGCTGAAC-1 
#>        0.5040557        0.5067164        0.5110906        0.5131760 
#> TCAGGATGCATGAC-1 TGTCTAACAAAACG-1 TTGGAGACCTAAGC-1 CTAATAGAACGGGA-1 
#>        0.5169279        0.5174344        0.5188547        0.5188547 
#> 
#> 
#> $`sample-1`[[1]]$`Cluster-5-dist1`
#> $`sample-1`[[1]]$`Cluster-5-dist1`[[1]]
#> ATGATATGAAGAGT-1 GAGGGAACTGGCAT-1 TGAACCGATCACGA-1 TGTTACTGTAACGC-1 
#>        0.4866294        0.5296660        0.5347212        0.5378081 
#> TGTCTAACAAAACG-1 ACCCACTGACACGT-1 GCATTGGAAAGGGC-1 TTGGAGACCTAAGC-1 
#>        0.5524990        0.5546457        0.5756563        0.5786301 
#> CGATACGAGTAGGG-1 ACTCCCGATACTCT-1 CGATCCACTGGCAT-1 TAGGTCGAACCCTC-1 
#>        0.5792414        0.5793405        0.5799661        0.5871738 
#> TAATCCACCTCAGA-1 CGAAGACTACCATG-1 CGGACCGACTCCAC-1 AGTGAAGAGAGAGC-1 
#>        0.5918016        0.5956331        0.5972017        0.5974153 
#> ACACGTGAAGTCAC-1 ACAAGAGATCGACA-1 ATTTCTCTGGACAG-1 ACAGTTCTTACGCA-1 
#>        0.5985295        0.6024728        0.6053899        0.6061259 
#> 
#> 
#> 
#> $`sample-1`[[2]]
#> $`sample-1`[[2]]$`Cluster-1-dist2`
#> $`sample-1`[[2]]$`Cluster-1-dist2`[[1]]
#> ATCGGAACTACGCA-1 AAACATTGTTTGTC-1 AAGAAGACTGACTG-1 ACTCCTCTCCACAA-1 
#>        0.5113990        0.5299588        0.5532387        0.5602466 
#> TGCTGAGATGAGCT-1 CCAGCGGATCCTAT-1 ACAGTGTGGGTGAG-1 ACTTAAGAACTGTG-1 
#>        0.5715726        0.5814976        0.5820154        0.5839420 
#> TGACACGACTGACA-1 ACCACAGACCTTGC-1 GTATTCACCAGGAG-1 CAGTGATGGGGTGA-1 
#>        0.5925851        0.5930390        0.5938605        0.5942740 
#> CGAGAACTCTGTAG-1 ACTACTACGTGTAC-1 TTCCCACTTTGAGC-1 ATATACGACCTCCA-1 
#>        0.5950798        0.5979864        0.6009199        0.6025974 
#> CCAGTGCTGTGCTA-1 CCATAGGACTTATC-1 ACCCGTTGGCTTCC-1 GTGTGATGGTTGAC-1 
#>        0.6039475        0.6051038        0.6054349        0.6089382 
#> 
#> 
#> $`sample-1`[[2]]$`Cluster-2-dist2`
#> $`sample-1`[[2]]$`Cluster-2-dist2`[[1]]
#> CAAGGTTGAGATCC-1 GACTACGATCTGGA-1 CGTGATGAATAAGG-1 AGCGAACTGTACAC-1 
#>        0.5746774        0.5866038        0.5872881        0.5881014 
#> ACTCGAGAACGGGA-1 CAAGCATGGACAAA-1 ATTCTTCTAGAAGT-1 CAGCTCACAACGAA-1 
#>        0.5926345        0.5951485        0.5957305        0.5980949 
#> TAAGAACTTAGAGA-1 AATATCGACCTACC-1 GAGTGACTTGGAGG-1 GTGACAACTACGAC-1 
#>        0.6029451        0.6043415        0.6056457        0.6123332 
#> GAATGCACCACCAA-1 AGTTTGCTATCACG-1 GAGTACTGCAAAGA-1 GTTATGCTTGGATC-1 
#>        0.6130043        0.6135373        0.6135607        0.6179548 
#> GATAATACTGGCAT-1 AAGATGGAACCACA-1 TGAATAACGAATGA-1 ATACGGACCATGGT-1 
#>        0.6183508        0.6187562        0.6192011        0.6220851 
#> 
#> 
#> $`sample-1`[[2]]$`Cluster-3-dist2`
#> $`sample-1`[[2]]$`Cluster-3-dist2`[[1]]
#>   GAAAGTGACGTACA-1   CCAGCACTCCATGA-1   GACCTAGACTTCCG-1   GTGGATTGTCTTAC-1 
#>          0.5809799          0.6102902          0.6154468          0.6221561 
#> GCTACGCTTGCCTC-1.1   AGCGCTCTGAGGCA-1   AATTGATGCTAGTG-1   ACTCGAGAACGGGA-1 
#>          0.6461629          0.6541032          0.6554948          0.6556701 
#>   GAATGCACCACCAA-1   TAGAATTGCCGCTT-1   GACAACACAGAGAT-1   CCGGAGACAGAGGC-1 
#>          0.6585368          0.6599428          0.6609825          0.6614554 
#>   TATCCTGAAGAATG-1   AGATATTGATGACC-1   ATTGCGGAGAAAGT-1   CGAGGCTGATTTCC-1 
#>          0.6652054          0.6666882          0.6700687          0.6702973 
#>   TAACCGGAACCCTC-1   TAGGAGCTCGTGTA-1   AGCACAACGCCAAT-1   GTCTGAGAGCCCTT-1 
#>          0.6721906          0.6728015          0.6750117          0.6754205 
#> 
#> 
#> $`sample-1`[[2]]$`Cluster-4-dist2`
#> $`sample-1`[[2]]$`Cluster-4-dist2`[[1]]
#> GAAAGTGACGTACA-1 ACTCCTCTCCACAA-1 CACTGCACCTCCAC-1 TAGAATTGCCGCTT-1 
#>        0.4502441        0.4590393        0.4610665        0.4679059 
#> ACTCGAGAACGGGA-1 CGAGGCTGATTTCC-1 CCAGCGGATCCTAT-1 TTCCCACTTTGAGC-1 
#>        0.4712896        0.4730695        0.4802572        0.4896431 
#> GCCACGGACATGAC-1 TAACCGGAACCCTC-1 CCGGAGACAGAGGC-1 GAGTACTGCAAAGA-1 
#>        0.4906466        0.4921628        0.4948145        0.4950318 
#> TGGTCAGATCCAAG-1 GTGGATTGTCTTAC-1 TACGGAACTGGCAT-1 TATCCTGAAGAATG-1 
#>        0.4965065        0.4966100        0.4970395        0.5017096 
#> TGAATAACGAATGA-1 CCCAGTTGCCCAAA-1 GAGTGACTTGGAGG-1 GACTACGATCTGGA-1 
#>        0.5050814        0.5054233        0.5064496        0.5069721 
#> 
#> 
#> $`sample-1`[[2]]$`Cluster-5-dist2`
#> $`sample-1`[[2]]$`Cluster-5-dist2`[[1]]
#> CCGGAGACAGAGGC-1 ACCACAGACCTTGC-1 CATAGTCTGACGAG-1 CCAGTGCTGGTGAG-1 
#>        0.5051263        0.5180117        0.5261178        0.5270345 
#> GTTAAAACGGAAAT-1 ACTCCTCTCCACAA-1 CGAGGCTGATTTCC-1 CGAGAACTGGAGGT-1 
#>        0.5277825        0.5493575        0.5525503        0.5547712 
#> CATGTTACGAATGA-1 GAGTACTGCAAAGA-1 TTCCCACTTTGAGC-1 CAGTGATGGGGTGA-1 
#>        0.5592944        0.5658408        0.5667726        0.5675981 
#> ACTTGACTCGCTAA-1 GAAAGTGACGTACA-1 GGGACCACTTTCAC-1 GCCACGGACATGAC-1 
#>        0.5682151        0.5686239        0.5692211        0.5701742 
#> TGAATAACGAATGA-1 AACACTCTACTAGC-1 AGGGCCACAACCGT-1 TGTGACGAATGCCA-1 
#>        0.5714708        0.5758827        0.5802708        0.5821705 
#> 
#> 
#> 
#> 
#> $`sample-2`
#> $`sample-2`[[1]]
#> $`sample-2`[[1]]$`Cluster-1-dist1`
#> $`sample-2`[[1]]$`Cluster-1-dist1`[[1]]
#> GCAGTTGATCAAGC-1 GCGAGAGAAGCCAT-1 CAATAATGGAAAGT-1 GTGTGATGGTTGAC-1 
#>        0.7200309        0.7252589        0.7636914        0.7776923 
#> CGAGGCTGATTTCC-1 TTCTAGTGGGAAAT-1 AAACATTGTTTGTC-1 ACTACTACGTGTAC-1 
#>        0.7785384        0.7851938        0.7912560        0.7917955 
#> CCAGTGCTGGTGAG-1 CCGGAGACAGAGGC-1 ACAGTGTGGGTGAG-1 ATTCGGGATCTTAC-1 
#>        0.7926018        0.7994835        0.8027161        0.8068640 
#> GTATTCACCAGGAG-1 ATAGCCGATTTGGG-1 TACGGAACTGGCAT-1 GACACTGACGTAGT-1 
#>        0.8074792        0.8097845        0.8111062        0.8147206 
#> CATACTACAACGAA-1 ACCACAGACCTTGC-1 AGGTTGTGCCTAAG-1 TGTGACGAATGCCA-1 
#>        0.8152215        0.8155384        0.8175348        0.8185208 
#> 
#> 
#> $`sample-2`[[1]]$`Cluster-2-dist1`
#> $`sample-2`[[1]]$`Cluster-2-dist1`[[1]]
#> TAGGCTGAGGACGA-1 ATTTCGTGAACAGA-1 ATCGGAACTACGCA-1 ACTCGAGACCTGTC-1 
#>        0.5693260        0.6312837        0.6490986        0.6622822 
#> ACTTGACTAACTGC-1 ACGTTTACCTTAGG-1 GGACCCGACACTGA-1 TTCTACGAGCGATT-1 
#>        0.6827076        0.6841784        0.6842953        0.6853387 
#> CCATAGGACTTATC-1 AGGACACTCAGATC-1 TGAACCGATGAGAA-1 CTTGAACTGAAACA-1 
#>        0.6865485        0.6937987        0.6988081        0.7076433 
#> AAGAAGACCTTGGA-1 TGCTGAGATGAGCT-1 AGAGGTCTCTTTAC-1 TGTTAAGAGTAGCT-1 
#>        0.7183231        0.7188018        0.7391025        0.7524326 
#> CTGAATCTGCTAAC-1 GGAGGATGGGACTT-1 GGAGGCCTTGAGAA-1 GGCAAGGACCCTAC-1 
#>        0.7536545        0.7791804        0.8091985        0.8107931 
#> 
#> 
#> $`sample-2`[[1]]$`Cluster-3-dist1`
#> $`sample-2`[[1]]$`Cluster-3-dist1`[[1]]
#> CGAGAACTGGAGGT-1 TGACACGACTGACA-1 GATCATCTTTCCAT-1 CCAGTGCTGGTGAG-1 
#>        0.5390199        0.5680987        0.5713032        0.5720847 
#> CCGGAGACAGAGGC-1 GTATTCACCAGGAG-1 GCGCACGAGAGATA-1 GGAGTTACTGCCAA-1 
#>        0.5825037        0.5851010        0.5995515        0.6152439 
#> CTTACATGTGGGAG-1 CGTTAACTGGGAGT-1 GACCATGACTGGAT-1 AAAGATCTGTTGCA-1 
#>        0.6167734        0.6172522        0.6215183        0.6232556 
#> CGTGATGAATAAGG-1 AGTTTGCTATCACG-1 CCAGTGCTGTGCTA-1 AGCGCTCTGAGGCA-1 
#>        0.6287034        0.6363177        0.6365964        0.6428942 
#> TTCTAGTGGGAAAT-1 CAGCTCACAACGAA-1 CATTGACTGTATGC-1 GGGAAGACCGTAGT-1 
#>        0.6545804        0.6556610        0.6580286        0.6618783 
#> 
#> 
#> $`sample-2`[[1]]$`Cluster-4-dist1`
#> $`sample-2`[[1]]$`Cluster-4-dist1`[[1]]
#>   CAGACATGACCTCC-1   ACCACAGACCTTGC-1   CGAGAACTCTGTAG-1   ACTACTACGTGTAC-1 
#>          0.4849671          0.4913897          0.5121517          0.5165632 
#>   CTCTAAACAGGAGC-1   CGAGGCTGATTTCC-1   ACAGACACTTTGGG-1   ACAGTGTGGGTGAG-1 
#>          0.5275871          0.5382640          0.5468898          0.5588711 
#>   GACACTGACGTAGT-1 GCTACGCTTGCCTC-1.1   TGTGACGAATGCCA-1   GAAAGTGACGTACA-1 
#>          0.5627945          0.5649240          0.5679256          0.5699943 
#>   TACGGAACTGGCAT-1   TAACCGGAACCCTC-1   GCAGTTGATCAAGC-1   CCACTGTGGCGAGA-1 
#>          0.5929589          0.6059879          0.6081365          0.6088404 
#>   TATAAGTGGTGTCA-1   AGACGTACCTACGA-1   CCAGCACTCCATGA-1   GACAACTGGTCGTA-1 
#>          0.6107916          0.6253095          0.6311408          0.6365964 
#> 
#> 
#> 
#> $`sample-2`[[2]]
#> $`sample-2`[[2]]$`Cluster-1-dist2`
#> $`sample-2`[[2]]$`Cluster-1-dist2`[[1]]
#> ATGCGCCTCTATGG-1 TCATGTACTGCTGA-1 TGTGAGTGCTTCGC-1 TTTCGAACACCCAA-1 
#>        0.6820170        0.6844507        0.6850731        0.6863869 
#> TGTAAAACGGTAGG-1 ACGCTGCTCAGCTA-1 TCGAGCCTGAGACG-1 ACATACCTAGAATG-1 
#>        0.6891738        0.6945091        0.6954656        0.7003688 
#> ATCTTGACCGCCTT-1 CAATAAACGCTGTA-1 CTTAGACTCTGTTT-1 CCCTGATGAGCGTT-1 
#>        0.7023863        0.7045970        0.7054086        0.7064281 
#> GCAGGCACGTACAC-1 AGCGAACTTGTAGC-1 CTCCATCTATACCG-1 ACTTCCCTTTTCGT-1 
#>        0.7110108        0.7111501        0.7112079        0.7121394 
#> CCCTCAGAGAGACG-1 ACCACAGACGATAC-1 ATCTTGACACCACA-1 CAGAGGGATACTGG-1 
#>        0.7135602        0.7141074        0.7163662        0.7196385 
#> 
#> 
#> $`sample-2`[[2]]$`Cluster-2-dist2`
#> $`sample-2`[[2]]$`Cluster-2-dist2`[[1]]
#> CGAGAACTGTTTGG-1 GTAAGCTGTGTCTT-1 GTGGATTGGCCAAT-1 TCCCGAACACGGAG-1 
#>        0.6542416        0.6580132        0.6712611        0.6758440 
#> AGCCTCTGGTACGT-1 GGCCAGACTTCCAT-1 TTCTTACTGGTACT-1 GCGATATGCTCATT-1 
#>        0.6788487        0.6874483        0.6884983        0.6952813 
#> AAAGACGACCAATG-1 CAGAGGGATACTGG-1 TCACCTCTCTATGG-1 CGGACCGACTCCAC-1 
#>        0.6991515        0.7007434        0.7042550        0.7090966 
#> GTCAATCTGGCAAG-1 TTGTCATGTGCACA-1 CAATTCTGGTCAAC-1 GGCGACTGCTCCAC-1 
#>        0.7103018        0.7150600        0.7159097        0.7207332 
#> CGTGTAGAAGACTC-1 GGTACAACGACTAC-1 GTAGCAACCACTTT-1 ATCTTGACCGCCTT-1 
#>        0.7226472        0.7272696        0.7287401        0.7336509 
#> 
#> 
#> $`sample-2`[[2]]$`Cluster-3-dist2`
#> $`sample-2`[[2]]$`Cluster-3-dist2`[[1]]
#> CTGCAGCTTCTTAC-1 GTGATCGAGTCTGA-1 AGTTCTACATTCTC-1 AGAGAATGTGACAC-1 
#>        0.5053621        0.5060096        0.5262206        0.5295006 
#> GATAGCACATAAGG-1 GGCCGAACAATCGC-1 ACCATTTGCATGCA-1 AACCTTACAGCGGA-1 
#>        0.5354795        0.5370609        0.5380423        0.5406374 
#> ACCTGAGACCTATT-1 TTCCCACTTTGAGC-1 CACGATGAGGCAAG-1 TACGCAGAGCCCTT-1 
#>        0.5421359        0.5427712        0.5447247        0.5530978 
#> ATACGGACAGCTAC-1 TGTAGTCTTCTACT-1 AAGATGGAACCACA-1 ATTCTTCTAGAAGT-1 
#>        0.5533231        0.5552665        0.5565558        0.5581426 
#> TGCAAGACCCCAAA-1 TAGGTTCTCAGTTG-1 AGACTGACCGTGAT-1 GAAACAGATTCATC-1 
#>        0.5600613        0.5609480        0.5615555        0.5639396 
#> 
#> 
#> $`sample-2`[[2]]$`Cluster-4-dist2`
#> $`sample-2`[[2]]$`Cluster-4-dist2`[[1]]
#> ACCTGAGACCTATT-1 ACTCCTCTCCACAA-1 CAGTGATGGGGTGA-1 CCAGCGGATCCTAT-1 
#>        0.4568545        0.4597093        0.4709658        0.4725519 
#> AGTTCTACGAGGTG-1 GAAACAGATTCATC-1 GGATGTTGAGCACT-1 AAATTCGATCAGTG-1 
#>        0.4791561        0.4822831        0.4842984        0.4876229 
#> CGCACTTGATTCTC-1 AAGATGGAACCACA-1 TACGTTACGGACGA-1 AAACGGCTTGTCAG-1 
#>        0.4886330        0.4895263        0.4944108        0.4956730 
#> ATTTGCACCTATGG-1 ACCCGTTGGCTTCC-1 CTTTGATGAAGGTA-1 GAACCAACCATGAC-1 
#>        0.4957854        0.4969039        0.4985343        0.4994797 
#> TCTAACTGCTGAAC-1 GGAGAGACCCTAAG-1 ATTAACGATCCAAG-1 CCACCTGAAGTCGT-1 
#>        0.5046081        0.5067114        0.5071328        0.5074389 
#> 
#> 
#> 
#> 
#> $`sample-3`
#> $`sample-3`[[1]]
#> $`sample-3`[[1]]$`Cluster-1-dist1`
#> $`sample-3`[[1]]$`Cluster-1-dist1`[[1]]
#> ACTTAAGAACTGTG-1 CTCCATCTATACCG-1 GTTATGCTTGGATC-1 AACCGATGTGCTTT-1 
#>        0.4819720        0.4886003        0.4971930        0.5044784 
#> ATATACGACCTCCA-1 GCCCAGGATTTCTG-1 GACGTAACCGCAAT-1 CCAGCGGATCCTAT-1 
#>        0.5046130        0.5353045        0.5370222        0.5466212 
#> AGGATAGACCGAAT-1 GCGTAAACCTCGAA-1 ACCCGTTGGCTTCC-1 GAGTGACTTGGAGG-1 
#>        0.5510425        0.5604996        0.5605754        0.5630457 
#> ACTCCTCTCCACAA-1 GTGATCGAGTCTGA-1 CTCGACTGATTTCC-1 TCAGTTACGGGAGT-1 
#>        0.5636650        0.5654089        0.5687629        0.5695756 
#> GGACCCGAATGCTG-1 CTGTAACTAGAACA-1 GAATGCACCACCAA-1 TGCAAGTGCCGAAT-1 
#>        0.5711248        0.5715321        0.5725934        0.5731254 
#> 
#> 
#> $`sample-3`[[1]]$`Cluster-2-dist1`
#> $`sample-3`[[1]]$`Cluster-2-dist1`[[1]]
#> CCTCTACTTTCTGT-1 TACTCTGACATGCA-1 CTCCACGACCGTAA-1 GTCGACCTACCTGA-1 
#>        0.5027937        0.5261415        0.5328501        0.5339133 
#> CATTGTTGTCACCC-1 TAATGATGCGCAAT-1 CTTGAGGATCAGAC-1 GTCTGAGAGCCCTT-1 
#>        0.5477846        0.5482708        0.5489740        0.5500259 
#> TGACCGCTAGTCAC-1 ACTCGAGAACGGGA-1 ACTCCCGAGTGTCA-1 CAATATGACGACAT-1 
#>        0.5572625        0.5592083        0.5600115        0.5669352 
#> ATCACTACAATCGC-1 TGCCCAACGCTACA-1 AGCGATACGCGAGA-1 AATTGATGCTAGTG-1 
#>        0.5671812        0.5790415        0.5810580        0.5838283 
#> GAATGCACCACCAA-1 GACTCCTGCCGTAA-1 AGACGTACCGGGAA-1 ATATACGACCTCCA-1 
#>        0.5840362        0.5881924        0.5904144        0.5913661 
#> 
#> 
#> $`sample-3`[[1]]$`Cluster-3-dist1`
#> $`sample-3`[[1]]$`Cluster-3-dist1`[[1]]
#> CTCGACTGATTTCC-1 GCAGCCGAGTCTGA-1 GCCCAGGATTTCTG-1 TTAACCACGCATAC-1 
#>        0.5058111        0.5182689        0.5272768        0.5353461 
#> TACAAATGTCAGTG-1 CGCTACTGAAGAAC-1 CCAGCGGATCCTAT-1 AATAGGGAGTTCGA-1 
#>        0.5368208        0.5430986        0.5438890        0.5455344 
#> CACCACTGCACTGA-1 TCGATACTCGAACT-1 ACCCGTTGGCTTCC-1 CTGTATACTCAGGT-1 
#>        0.5497586        0.5510013        0.5514001        0.5519565 
#> TGCAAGTGCCGAAT-1 ACGTTTACAGACTC-1 TCGCAGCTTCCCAC-1 CATGTTACGAATGA-1 
#>        0.5560525        0.5567003        0.5569072        0.5618946 
#> AGCGAACTTGTAGC-1 ACTTGACTCGCTAA-1 GACTCCTGCCGTAA-1 TGACTTACTTGCTT-1 
#>        0.5649894        0.5675069        0.5688738        0.5691047 
#> 
#> 
#> $`sample-3`[[1]]$`Cluster-4-dist1`
#> $`sample-3`[[1]]$`Cluster-4-dist1`[[1]]
#> GGAGAGACCCTAAG-1 ACCCGTTGGCTTCC-1 GCCACGGACATGAC-1 AGCGAACTTGTAGC-1 
#>        0.4659939        0.4790934        0.5061603        0.5099404 
#> AATTGATGCTAGTG-1 TTCCCACTTTGAGC-1 ATATACGACCTCCA-1 CTTGAGGATCAGAC-1 
#>        0.5184120        0.5250526        0.5257730        0.5287571 
#> TCGCAGCTTCCCAC-1 CTGAACGATACGCA-1 AAGATGGAACCACA-1 GCCCAGGATTTCTG-1 
#>        0.5309801        0.5335436        0.5382103        0.5382443 
#> CTCGACTGATTTCC-1 GATAATACTGGCAT-1 GCGTAAACCTCGAA-1 ACTTAAGAACTGTG-1 
#>        0.5405251        0.5409792        0.5452960        0.5471226 
#> AATAGGGAGTTCGA-1 ATTCTTCTAGAAGT-1 GCTATACTCATGCA-1 CATCCCGACACTCC-1 
#>        0.5521107        0.5521162        0.5550496        0.5551539 
#> 
#> 
#> $`sample-3`[[1]]$`Cluster-5-dist1`
#> $`sample-3`[[1]]$`Cluster-5-dist1`[[1]]
#> CTCCACGACCGTAA-1 CCAAGTGATGACTG-1 AGGGTTTGTCTTCA-1 GTTATGCTTGGATC-1 
#>        0.6250397        0.6427113        0.6437552        0.6629978 
#> GTATTCACCATCAG-1 GGACTATGCCGCTT-1 AAGTGCACCACAAC-1 ATGGGTACGGGACA-1 
#>        0.6664160        0.6681879        0.6742650        0.6746051 
#> ACGAGGGAGTTGTG-1 GCAGGGCTAACCAC-1 TCTAACTGTGTCCC-1 ACTTGACTCGCTAA-1 
#>        0.6773157        0.6809019        0.6814398        0.6819156 
#> AGTTTAGATCTCGC-1 GGGAAGTGCCACAA-1 ATTGCGGAGAAAGT-1 ACGTCAGAGGTCAT-1 
#>        0.6838746        0.6845133        0.6864370        0.6865782 
#> AGGCAACTGAGATA-1 TAAACAACCTCGCT-1 CATCCCGAGGATCT-1 CTTGAGGATCAGAC-1 
#>        0.6867028        0.6882336        0.6919740        0.6920532 
#> 
#> 
#> 
#> $`sample-3`[[2]]
#> $`sample-3`[[2]]$`Cluster-1-dist2`
#> $`sample-3`[[2]]$`Cluster-1-dist2`[[1]]
#> TACTCTGAGTCTTT-1 TGTAGTCTTCTACT-1 CTTAGACTCTGTTT-1 CACCGTACCTGATG-1 
#>        0.4508019        0.4903765        0.4921710        0.5108653 
#> TGACGCCTACACCA-1 AATGAGGAGGTAGG-1 TGTAAAACGGTAGG-1 CGTGATGAATAAGG-1 
#>        0.5141737        0.5146790        0.5154864        0.5163589 
#> AATAGGGAGAGCTT-1 AACCGCCTTAAAGG-1 GCCAACCTTTTCAC-1 ACTTCCCTTTTCGT-1 
#>        0.5254959        0.5278003        0.5281624        0.5301507 
#> ACACGTGAAGTCAC-1 CGTCCATGCAGAAA-1 TCATGTACTGCTGA-1 CTGCAGCTTCTTAC-1 
#>        0.5306401        0.5306839        0.5309892        0.5316392 
#> TATACAGATTGACG-1 AAATTCGATCAGTG-1 CGATCCACTGGCAT-1 CTGCGACTTTGAGC-1 
#>        0.5334844        0.5357404        0.5393725        0.5431064 
#> 
#> 
#> $`sample-3`[[2]]$`Cluster-2-dist2`
#> $`sample-3`[[2]]$`Cluster-2-dist2`[[1]]
#> CCCGGAGATTGAGC-1 ACCGAAACTGACCA-1 CGGATAACCACAAC-1 CGTCCATGCAGAAA-1 
#>        0.5269517        0.5395385        0.5422591        0.5435326 
#> CACCGTACCTGATG-1 TAATCCACCTCAGA-1 AATCTCTGCGAATC-1 ATTGAATGTCGTGA-1 
#>        0.5542126        0.5562237        0.5577426        0.5618496 
#> ACACGTGAAGTCAC-1 GTCAATCTTCAGTG-1 TCGGCACTCCAAGT-1 ACCATTTGCATGCA-1 
#>        0.5625340        0.5648604        0.5685800        0.5690515 
#> ACTATCACCTCTAT-1 TAAGCGTGCGATAC-1 ATATAGTGTAAGCC-1 TACTCTGAGTCTTT-1 
#>        0.5746944        0.5754268        0.5819473        0.5826056 
#> GTTGATCTTGGAGG-1 TATAGATGCGGGAA-1 GTTAAATGCTGTTT-1 TGTCTAACAAAACG-1 
#>        0.5849699        0.5898461        0.5906179        0.5921320 
#> 
#> 
#> $`sample-3`[[2]]$`Cluster-3-dist2`
#> $`sample-3`[[2]]$`Cluster-3-dist2`[[1]]
#> CGATCCACTGGCAT-1 CTGCAGCTTCTTAC-1 GTCACAGAGAGACG-1 ACTTGACTCACCAA-1 
#>        0.5268830        0.5279582        0.5347179        0.5357977 
#> CACCGTACCTGATG-1 TACGCAGAGCCCTT-1 ACACGTGAAGTCAC-1 TACTCTGAGTCTTT-1 
#>        0.5496908        0.5548494        0.5592791        0.5620121 
#> ACCCACTGACACGT-1 GTGTGATGGTTGAC-1 CCGGAGACAGAGGC-1 TAACCGGAACCCTC-1 
#>        0.5629403        0.5722983        0.5724597        0.5759515 
#> CGAGAACTGGAGGT-1 AGCGGCTGCCGTTC-1 CTCCACGACAAGCT-1 CTAATAGAACGGGA-1 
#>        0.5797644        0.5816495        0.5821794        0.5838569 
#> CGCACTTGATTCTC-1 TATGTGCTAACGAA-1 GATCGTGAGACAAA-1 ACTATCACCTCTAT-1 
#>        0.5844430        0.5855951        0.5857692        0.5866885 
#> 
#> 
#> $`sample-3`[[2]]$`Cluster-4-dist2`
#> $`sample-3`[[2]]$`Cluster-4-dist2`[[1]]
#> ACTACTACGTGTAC-1 TGTAGTCTTCTACT-1 AATGAGGAGGTAGG-1 CTATTGTGGGTACT-1 
#>        0.4507327        0.4607078        0.4618565        0.4641227 
#> CGTGATGAATAAGG-1 ACCTGAGACCTATT-1 TACTCTGAGTCTTT-1 AACCGCCTTAAAGG-1 
#>        0.4696448        0.4731942        0.4768857        0.4772024 
#> AACCTTACAGCGGA-1 CACCGTACCTGATG-1 GCGCACGAGAGATA-1 TGACACGACTGACA-1 
#>        0.4804801        0.4814813        0.4859867        0.4865429 
#> CTGCAGCTTCTTAC-1 CGCACTTGATTCTC-1 ACCACAGACCTTGC-1 CGATCCACTGGCAT-1 
#>        0.4892227        0.4900210        0.4955669        0.4957772 
#> CGAGAACTCTGTAG-1 CTGCGACTTTGAGC-1 CTACGCACTACGAC-1 AAACGGCTTGTCAG-1 
#>        0.4998494        0.5028970        0.5043626        0.5064802 
#> 
#> 
#> $`sample-3`[[2]]$`Cluster-5-dist2`
#> $`sample-3`[[2]]$`Cluster-5-dist2`[[1]]
#> CGTACCACCTGCTC-1 TAACCGGAACCCTC-1 CTATTGACTCCGTC-1 TATAAGTGGTGTCA-1 
#>        0.6371825        0.6465647        0.6490919        0.6715657 
#> TAAGCGTGCGATAC-1 ACACGTGAAGTCAC-1 GAAAGTGACGTACA-1 GATTTAGAATCTTC-1 
#>        0.6790751        0.6808244        0.6839177        0.6864370 
#> TCCATAACAAGGCG-1 ACTATCACCACTAG-1 CACCGTACCTGATG-1 ACAAGAGATCGACA-1 
#>        0.6917816        0.6919323        0.6922012        0.6971436 
#> CCAGCACTCCATGA-1 AGCGGCTGCCGTTC-1 ACGTAGACCTCTAT-1 ATTGAATGTCGTGA-1 
#>        0.6978365        0.7060329        0.7071385        0.7102035 
#> AGTGAAGAGAGAGC-1 GTCACAGAGAGACG-1 CTCGCATGAGAATG-1 CGATCCACTGGCAT-1 
#>        0.7126484        0.7181470        0.7184851        0.7191582 
#> 
#> 
#> 
#>