Get broad and fine clusters

Get broad and fine clusters

GetCluster(seuratlist, n1 = 50, n2 = 200)

Arguments

seuratlist

A list of Seurat objects, usually can be got by SplitObject(). We also accept the SingleCellExperiment object input.

n1

If the number of cells was smaller than n1, then the cluster will remain unchanged called rare cluster. The default value of n1 is 50.

n2

If the count of cells within a broad cluster is more than n2, the cluster is subdivided randomly into three fine clusters. If the cell count falls within the range of n1 to n2, two fine clusters are generated randomly. Default value is 200.

Value

A list of data frames.

Examples

data(sim_data_sce)
# Assuming "seuratlist" is a list of Seurat objects
# if(is(sim_data_sce, "SingleCellExperiment")){ sim_data <- as.Seurat(sim_data_sce) }
sim_data <- SCEtoSeurat(sim_data_sce)
#> Warning: No gene metadata columns found to add to Seurat object.
seuratlist <- Seurat::SplitObject(sim_data, split.by = "Study")
# seuratlist <- InputData(sim_data_sce,"Study")
fullcluster <- GetCluster(seuratlist)
#> Normalizing layer: counts
#> Finding variable features for layer counts
#> Warning: pseudoinverse used at -2.5443
#> Warning: neighborhood radius 0.31902
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.090619
#> Centering and scaling data matrix
#> Warning: The following 1324 features requested have zero variance; running reduction without them: ENSG00000223865	HLA-DPB1, ENSG00000160888	IER2, ENSG00000181035	SLC25A42, ENSG00000125726	CD70, ENSG00000099956	SMARCB1, ENSG00000167196	FBXO22, ENSG00000198286	CARD11, ENSG00000198355	PIM3, ENSG00000134460	IL2RA, ENSG00000115233	PSMD14, ENSG00000163156	SCNM1, ENSG00000102575	ACP5, ENSG00000165915	SLC39A13, ENSG00000101194	SLC17A9, ENSG00000214900	C14orf182, ENSG00000198821	CD247, ENSG00000139193	CD27, ENSG00000219545	RPA3-AS1, ENSG00000256060	TRAPPC2P1, ENSG00000146143	PRIM2, ENSG00000135913	USP37, ENSG00000272686	RP11-390E23.6, ENSG00000136152	COG3, ENSG00000103091	WDR59, ENSG00000137411	VARS2, ENSG00000108349	CASC3, ENSG00000063761	ADCK1, ENSG00000197535	MYO5A, ENSG00000182685	BRICD5, ENSG00000145908	ZNF300, ENSG00000090989	EXOC1, ENSG00000126759	CFP, ENSG00000268006	PTOV1-AS1, ENSG00000168434	COG7, ENSG00000111269	CREBL2, ENSG00000174456	C12orf76, ENSG00000184674	GSTT1, ENSG00000156170	NDUFAF6, ENSG00000167642	SPINT2, ENSG00000133606	MKRN1, ENSG00000149182	ARFGAP2, ENSG00000169100	SLC25A6, ENSG00000112799	LY86, ENSG00000171148	TADA3, ENSG00000101220	C20orf27, ENSG00000204252	HLA-DOA, ENSG00000213977	TAX1BP3, ENSG00000127774	EMC6, ENSG00000120868	APAF1, ENSG00000173064	HECTD4, ENSG00000156973	PDE6D, ENSG00000088766	CRLS1, ENSG00000135521	LTV1, ENSG00000129158	SERGEF, ENSG00000105443	CYTH2, ENSG00000090621	PABPC4, ENSG00000162086	ZNF75A, ENSG00000105341	ATP5SL, ENSG00000084072	PPIE, ENSG00000100983	GSS, ENSG00000100908	EMC9, ENSG00000198521	ZNF43, ENSG00000138468	SENP7, ENSG00000140263	SORD, ENSG00000175155	YPEL2, ENSG00000156599	ZDHHC5, ENSG00000204001	LCN8, ENSG00000148719	DNAJB12, ENSG00000137193	PIM1, ENSG00000138138	ATAD1, ENSG00000129932	DOHH, ENSG00000178691	SUZ12, ENSG00000117335	CD46, ENSG00000172216	CEBPB, ENSG00000174720	LARP7, ENSG00000139725	RHOF, ENSG00000120742	SERP1, ENSG00000185507	IRF7, ENSG00000130787	HIP1R, ENSG00000122971	ACADS, ENSG00000173264	GPR137, ENSG00000141101	NOB1, ENSG00000080189	SLC35C2, ENSG00000143553	SNAPIN, ENSG00000166228	PCBD1, ENSG00000109534	GAR1, ENSG00000204713	TRIM27, ENSG00000100811	YY1, ENSG00000128272	ATF4, ENSG00000072694	FCGR2B, ENSG00000148300	REXO4, ENSG00000012779	ALOX5, ENSG00000099992	TBC1D10A, ENSG00000116497	S100PBP, ENSG00000132950	ZMYM5, ENSG00000143458	GABPB2, ENSG00000114416	FXR1, ENSG00000105993	DNAJB6, ENSG00000139921	TMX1, ENSG00000091009	RBM27, ENSG00000189227	C15orf61, ENSG00000167766	ZNF83, ENSG00000107771	CCSER2, ENSG00000173933	RBM4, ENSG00000131013	PPIL4, ENSG00000055163	CYFIP2, ENSG00000168701	TMEM208, ENSG00000124214	STAU1, ENSG00000223482	NUTM2A-AS1, ENSG00000164733	CTSB, ENSG00000146729	GBAS, ENSG00000114503	NCBP2, ENSG00000145416	MARCH1, ENSG00000169249	ZRSR2, ENSG00000073614	KDM5A, ENSG00000181396	OGFOD3, ENSG00000148248	SURF4, ENSG00000173221	GLRX, ENSG00000118680	MYL12B, ENSG00000100359	SGSM3, ENSG00000168209	DDIT4, ENSG00000213240	NOTCH2NL, ENSG00000114120	SLC25A36, ENSG00000099246	RAB18, ENSG00000167862	ICT1, ENSG00000100949	RABGGTA, ENSG00000072364	AFF4, ENSG00000157456	CCNB2, ENSG00000167513	CDT1, ENSG00000169981	ZNF35, ENSG00000069275	NUCKS1, ENSG00000167815	PRDX2, ENSG00000213246	SUPT4H1, ENSG00000196230	TUBB, ENSG00000204231	RXRB, ENSG00000169032	MAP2K1, ENSG00000272831	RP11-792A8.4, ENSG00000170248	PDCD6IP, ENSG00000162642	C1orf52, ENSG00000197223	C1D, ENSG00000121931	LRIF1, ENSG00000140307	GTF2A2, ENSG00000174529	TMEM81, ENSG00000187266	EPOR, ENSG00000169679	BUB1, ENSG00000132382	MYBBP1A, ENSG00000185305	ARL15, ENSG00000214960	ISPD, ENSG00000230424	RP1-43E13.2, ENSG00000175906	ARL4D, ENSG00000104490	NCALD, ENSG00000255062	RP11-712L6.5, ENSG00000115904	SOS1, ENSG00000197937	ZNF347, ENSG00000139514	SLC7A1, ENSG00000077458	FAM76B, ENSG00000130684	ZNF337, ENSG00000152969	JAKMIP1, ENSG00000155324	GRAMD3, ENSG00000178163	ZNF518B, ENSG00000254415	SIGLEC14, ENSG00000235192	AC009495.2, ENSG00000184451	CCR10, ENSG00000129317	PUS7L, ENSG00000102967	DHODH, ENSG00000141424	SLC39A6, ENSG00000268364	SMC5-AS1, ENSG00000261105	RP11-173B14.5, ENSG00000010379	SLC6A13, ENSG00000264078	RP11-73M7.9, ENSG00000236933	RP11-439A17.7, ENSG00000115368	WDR75, ENSG00000153767	GTF2E1, ENSG00000132561	MATN2, ENSG00000255337	RP11-315O6.1, ENSG00000143167	GPA33, ENSG00000184708	EIF4ENIF1, ENSG00000237596	RP13-143G15.4, ENSG00000131759	RARA, ENSG00000188707	ZBED6CL, ENSG00000198742	SMURF1, ENSG00000163002	NUP35, ENSG00000140525	FANCI, ENSG00000196792	STRN3, ENSG00000232995	RP11-267N12.3, ENSG00000168994	PXDC1, ENSG00000213967	ZNF726, ENSG00000158158	CNNM4, ENSG00000224597	PTCHD3P1, ENSG00000113742	CPEB4, ENSG00000140545	MFGE8, ENSG00000113812	ACTR8, ENSG00000253210	RP11-809O17.1, ENSG00000158828	PINK1, ENSG00000188157	AGRN, ENSG00000085563	ABCB1, ENSG00000135124	P2RX4, ENSG00000243244	STON1, ENSG00000147874	HAUS6, ENSG00000115159	GPD2, ENSG00000125247	TMTC4, ENSG00000112039	FANCE, ENSG00000133466	C1QTNF6, ENSG00000183647	ZNF530, ENSG00000247199	RP11-373N22.3, ENSG00000092470	WDR76, ENSG00000180089	TMEM86B, ENSG00000116199	FAM20B, ENSG00000152253	SPC25, ENSG00000116273	PHF13, ENSG00000010310	GIPR, ENSG00000114573	ATP6V1A, ENSG00000166484	MAPK7, ENSG00000136231	IGF2BP3, ENSG00000272986	RP11-46J23.1, ENSG00000140950	TLDC1, ENSG00000187626	ZKSCAN4, ENSG00000160113	NR2F6, ENSG00000103222	ABCC1, ENSG00000229729	RP11-159G9.5, ENSG00000223759	RP11-353N4.4, ENSG00000162378	ZYG11B, ENSG00000198517	MAFK, ENSG00000196739	COL27A1, ENSG00000103260	METRN, ENSG00000152582	SPEF2, ENSG00000245146	LINC01024, ENSG00000111252	SH2B3, ENSG00000162616	DNAJB4, ENSG00000141198	TOM1L1, ENSG00000185291	IL3RA, ENSG00000109184	DCUN1D4, ENSG00000235831	BHLHE40-AS1, ENSG00000135966	TGFBRAP1, ENSG00000167536	DHRS13, ENSG00000009830	POMT2, ENSG00000163067	ZNF2, ENSG00000166963	MAP1A, ENSG00000155363	MOV10, ENSG00000176225	RTTN, ENSG00000152705	CATSPER3, ENSG00000075234	TTC38, ENSG00000173276	ZBTB21, ENSG00000171823	FBXL14, ENSG00000138160	KIF11, ENSG00000197256	KANK2, ENSG00000091106	NLRC4, ENSG00000213214	ARHGEF35, ENSG00000161021	MAML1, ENSG00000155252	PI4K2A, ENSG00000247796	CTD-2366F13.1, ENSG00000154447	SH3RF1, ENSG00000228434	AC004951.6, ENSG00000080200	CRYBG3, ENSG00000153898	MCOLN2, ENSG00000139344	AMDHD1, ENSG00000204788	CR769776.1, ENSG00000182508	LHFPL1, ENSG00000166529	ZSCAN21, ENSG00000181284	TMEM102, ENSG00000227258	SMIM2-AS1, ENSG00000184007	PTP4A2, ENSG00000213676	ATF6B, ENSG00000183431	SF3A3, ENSG00000145687	SSBP2, ENSG00000079246	XRCC5, ENSG00000118705	RPN2, ENSG00000187257	RSBN1L, ENSG00000149311	ATM, ENSG00000153774	CFDP1, ENSG00000067334	DNTTIP2, ENSG00000166260	COX11, ENSG00000134884	ARGLU1, ENSG00000145741	BTF3, ENSG00000115053	NCL, ENSG00000143575	HAX1, ENSG00000084234	APLP2, ENSG00000166333	ILK, ENSG00000169288	MRPL1, ENSG00000167578	RAB4B, ENSG00000150347	ARID5B, ENSG00000172590	MRPL52, ENSG00000108774	RAB5C, ENSG00000129534	MIS18BP1, ENSG00000108010	GLRX3, ENSG00000241839	PLEKHO2, ENSG00000169398	PTK2, ENSG00000178199	ZC3H12D, ENSG00000175322	ZNF519, ENSG00000179134	SAMD4B, ENSG00000112640	PPP2R5D, ENSG00000263335	AF001548.5, ENSG00000186376	ZNF75D, ENSG00000166133	RPUSD2, ENSG00000273314	RP5-1136G13.2, ENSG00000166965	RCCD1, ENSG00000231889	TRAF3IP2-AS1, ENSG00000114450	GNB4, ENSG00000159753	RLTPR, ENSG00000130714	POMT1, ENSG00000049167	ERCC8, ENSG00000132953	XPO4, ENSG00000154845	PPP4R1, ENSG00000197948	FCHSD1, ENSG00000118292	C1orf54, ENSG00000107951	MTPAP, ENSG00000273456	RP11-686O6.2, ENSG00000253426	RP11-10A14.4, ENSG00000204138	PHACTR4, ENSG00000165359	DDX26B, ENSG00000129472	RAB2B, ENSG00000261079	RP11-252A24.3, ENSG00000149657	LSM14B, ENSG00000254470	AP5B1, ENSG00000163349	HIPK1, ENSG00000267519	MIR24-2, ENSG00000185716	C16orf52, ENSG00000105088	OLFM2, ENSG00000197183	C20orf112, ENSG00000257303	RP11-977G19.11, ENSG00000137494	ANKRD42, ENSG00000213380	COG8, ENSG00000237438	CECR7, ENSG00000106003	LFNG, ENSG00000153140	CETN3, ENSG00000139190	VAMP1, ENSG00000198920	KIAA0753, ENSG00000021574	SPAST, ENSG00000204315	FKBPL, ENSG00000173442	EHBP1L1, ENSG00000181847	TIGIT, ENSG00000184154	LRTOMT, ENSG00000166979	EVA1C, ENSG00000241978	AKAP2, ENSG00000051382	PIK3CB, ENSG00000165675	ENOX2, ENSG00000197841
#> PC_ 1 
#> Positive:  ENSG00000063177	RPL18, ENSG00000162244	RPL29, ENSG00000105193	RPS16, ENSG00000118518	RNF146, ENSG00000175787	ZNF169, ENSG00000257698	RP11-620J15.3, ENSG00000142784	WDTC1, ENSG00000229474	PATL2, ENSG00000266967	AARSD1, ENSG00000262202	SNORD3D 
#> 	   ENSG00000167483	FAM129C, ENSG00000135587	SMPD2, ENSG00000260000	RP3-467N11.1, ENSG00000197989	SNHG12, ENSG00000164077	MON1A, ENSG00000114631	PODXL2, ENSG00000085982	USP40, ENSG00000156515	HK1, ENSG00000204482	LST1, ENSG00000186166	CCDC84 
#> 	   ENSG00000271858	XXcos-LUCA11.4, ENSG00000273344	PAXIP1-AS1, ENSG00000105866	SP4, ENSG00000003249	DBNDD1, ENSG00000102974	CTCF, ENSG00000163960	UBXN7, ENSG00000102409	BEX4, ENSG00000146757	ZNF92, ENSG00000186265	BTLA, ENSG00000166145	SPINT1 
#> Negative:  ENSG00000166387	PPFIBP2, ENSG00000123240	OPTN, ENSG00000082641	NFE2L1, ENSG00000197157	SND1, ENSG00000030582	GRN, ENSG00000143228	NUF2, ENSG00000161980	POLR3K, ENSG00000158481	CD1C, ENSG00000111206	FOXM1, ENSG00000072571	HMMR 
#> 	   ENSG00000198546	ZNF511, ENSG00000213523	SRA1, ENSG00000108826	MRPL27, ENSG00000214756	METTL12, ENSG00000148950	IMMP1L, ENSG00000238227	C9orf69, ENSG00000143933	CALM2, ENSG00000215021	PHB2, ENSG00000115128	SF3B14, ENSG00000156411	C14orf2 
#> 	   ENSG00000172366	FAM195A, ENSG00000026025	VIM, ENSG00000197043	ANXA6, ENSG00000080822	CLDND1, ENSG00000090554	FLT3LG, ENSG00000088986	DYNLL1, ENSG00000173559	NABP1, ENSG00000127838	PNKD, ENSG00000116030	SUMO1, ENSG00000174886	NDUFA11 
#> PC_ 2 
#> Positive:  ENSG00000090554	FLT3LG, ENSG00000126368	NR1D1, ENSG00000120915	EPHX2, ENSG00000020256	ZFP64, ENSG00000269089	AP003733.1, ENSG00000139197	PEX5, ENSG00000178773	CPNE7, ENSG00000150054	MPP7, ENSG00000229539	RP11-119B16.2, ENSG00000171302	CANT1 
#> 	   ENSG00000160087	UBE2J2, ENSG00000159496	RGL4, ENSG00000166971	AKTIP, ENSG00000155366	RHOC, ENSG00000005001	PRSS22, ENSG00000273247	RP11-83A24.2, ENSG00000148950	IMMP1L, ENSG00000213523	SRA1, ENSG00000080822	CLDND1, ENSG00000198546	ZNF511 
#> 	   ENSG00000110330	BIRC2, ENSG00000105829	BET1, ENSG00000120948	TARDBP, ENSG00000155438	NIFK, ENSG00000170175	CHRNB1, ENSG00000111348	ARHGDIB, ENSG00000196642	RABL6, ENSG00000198130	HIBCH, ENSG00000015479	MATR3, ENSG00000163960	UBXN7 
#> Negative:  ENSG00000158481	CD1C, ENSG00000239961	LILRA4, ENSG00000072518	MARK2, ENSG00000161091	MFSD12, ENSG00000107581	EIF3A, ENSG00000198715	C1orf85, ENSG00000173559	NABP1, ENSG00000215021	PHB2, ENSG00000198700	IPO9, ENSG00000030582	GRN 
#> 	   ENSG00000178971	CTC1, ENSG00000196954	CASP4, ENSG00000063177	RPL18, ENSG00000159314	ARHGAP27, ENSG00000167522	ANKRD11, ENSG00000160326	SLC2A6, ENSG00000020633	RUNX3, ENSG00000066336	SPI1, ENSG00000127527	EPS15L1, ENSG00000172366	FAM195A 
#> 	   ENSG00000170759	KIF5B, ENSG00000123240	OPTN, ENSG00000143933	CALM2, ENSG00000166595	FAM96B, ENSG00000095059	DHPS, ENSG00000176619	LMNB2, ENSG00000039650	PNKP, ENSG00000196262	PPIA, ENSG00000164088	PPM1M, ENSG00000169905	TOR1AIP2 
#> PC_ 3 
#> Positive:  ENSG00000170175	CHRNB1, ENSG00000168303	MPLKIP, ENSG00000126368	NR1D1, ENSG00000239961	LILRA4, ENSG00000120915	EPHX2, ENSG00000198546	ZNF511, ENSG00000161091	MFSD12, ENSG00000172366	FAM195A, ENSG00000158481	CD1C, ENSG00000171302	CANT1 
#> 	   ENSG00000107581	EIF3A, ENSG00000095059	DHPS, ENSG00000020256	ZFP64, ENSG00000125637	PSD4, ENSG00000135916	ITM2C, ENSG00000257704	PRR24, ENSG00000072518	MARK2, ENSG00000269089	AP003733.1, ENSG00000110330	BIRC2, ENSG00000176619	LMNB2 
#> 	   ENSG00000243279	PRAF2, ENSG00000090447	TFAP4, ENSG00000164077	MON1A, ENSG00000173559	NABP1, ENSG00000266964	FXYD1, ENSG00000229539	RP11-119B16.2, ENSG00000178773	CPNE7, ENSG00000150054	MPP7, ENSG00000139197	PEX5, ENSG00000039650	PNKP 
#> Negative:  ENSG00000170540	ARL6IP1, ENSG00000273080	RP11-301O19.1, ENSG00000187118	CMC1, ENSG00000166387	PPFIBP2, ENSG00000131378	RFTN1, ENSG00000137133	HINT2, ENSG00000161980	POLR3K, ENSG00000136527	TRA2B, ENSG00000196642	RABL6, ENSG00000163960	UBXN7 
#> 	   ENSG00000159496	RGL4, ENSG00000121064	SCPEP1, ENSG00000236104	ZBTB22, ENSG00000089123	TASP1, ENSG00000082641	NFE2L1, ENSG00000153896	ZNF599, ENSG00000128191	DGCR8, ENSG00000102409	BEX4, ENSG00000155330	C16orf87, ENSG00000156411	C14orf2 
#> 	   ENSG00000136717	BIN1, ENSG00000233937	CTC-338M12.4, ENSG00000154473	BUB3, ENSG00000225828	FAM229A, ENSG00000197157	SND1, ENSG00000165934	CPSF2, ENSG00000170606	HSPA4, ENSG00000197989	SNHG12, ENSG00000003249	DBNDD1, ENSG00000162620	LRRIQ3 
#> PC_ 4 
#> Positive:  ENSG00000105619	TFPT, ENSG00000143933	CALM2, ENSG00000167543	TP53I13, ENSG00000125743	SNRPD2, ENSG00000198715	C1orf85, ENSG00000159496	RGL4, ENSG00000020633	RUNX3, ENSG00000109046	WSB1, ENSG00000118518	RNF146, ENSG00000162076	FLYWCH2 
#> 	   ENSG00000167522	ANKRD11, ENSG00000166595	FAM96B, ENSG00000090554	FLT3LG, ENSG00000072518	MARK2, ENSG00000273080	RP11-301O19.1, ENSG00000158481	CD1C, ENSG00000137404	NRM, ENSG00000233937	CTC-338M12.4, ENSG00000089123	TASP1, ENSG00000187118	CMC1 
#> 	   ENSG00000162620	LRRIQ3, ENSG00000121064	SCPEP1, ENSG00000168906	MAT2A, ENSG00000257704	PRR24, ENSG00000178971	CTC1, ENSG00000117505	DR1, ENSG00000108591	DRG2, ENSG00000162928	PEX13, ENSG00000102309	PIN4, ENSG00000116030	SUMO1 
#> Negative:  ENSG00000143228	NUF2, ENSG00000082641	NFE2L1, ENSG00000148840	PPRC1, ENSG00000160741	CRTC2, ENSG00000111224	PARP11, ENSG00000166387	PPFIBP2, ENSG00000128039	SRD5A3, ENSG00000143353	LYPLAL1, ENSG00000238227	C9orf69, ENSG00000111206	FOXM1 
#> 	   ENSG00000072571	HMMR, ENSG00000165733	BMS1, ENSG00000247240	UBL7-AS1, ENSG00000136147	PHF11, ENSG00000214756	METTL12, ENSG00000080822	CLDND1, ENSG00000154473	BUB3, ENSG00000257698	RP11-620J15.3, ENSG00000165934	CPSF2, ENSG00000105193	RPS16 
#> 	   ENSG00000030582	GRN, ENSG00000184886	PIGW, ENSG00000160087	UBE2J2, ENSG00000092036	HAUS4, ENSG00000272606	RP11-554J4.1, ENSG00000141002	TCF25, ENSG00000266967	AARSD1, ENSG00000100564	PIGH, ENSG00000134882	UBAC2, ENSG00000197043	ANXA6 
#> PC_ 5 
#> Positive:  ENSG00000173559	NABP1, ENSG00000142856	ITGB3BP, ENSG00000020256	ZFP64, ENSG00000150776	C11orf57, ENSG00000120948	TARDBP, ENSG00000163960	UBXN7, ENSG00000167118	URM1, ENSG00000187118	CMC1, ENSG00000169905	TOR1AIP2, ENSG00000136819	C9orf78 
#> 	   ENSG00000159496	RGL4, ENSG00000214756	METTL12, ENSG00000159479	MED8, ENSG00000109919	MTCH2, ENSG00000101265	RASSF2, ENSG00000198715	C1orf85, ENSG00000167483	FAM129C, ENSG00000164327	RICTOR, ENSG00000224051	GLTPD1, ENSG00000161091	MFSD12 
#> 	   ENSG00000227124	ZNF717, ENSG00000239961	LILRA4, ENSG00000110330	BIRC2, ENSG00000160087	UBE2J2, ENSG00000170100	ZNF778, ENSG00000121060	TRIM25, ENSG00000105499	PLA2G4C, ENSG00000136147	PHF11, ENSG00000164088	PPM1M, ENSG00000145414	NAF1 
#> Negative:  ENSG00000198026	ZNF335, ENSG00000229539	RP11-119B16.2, ENSG00000260000	RP3-467N11.1, ENSG00000047932	GOPC, ENSG00000273344	PAXIP1-AS1, ENSG00000233937	CTC-338M12.4, ENSG00000198700	IPO9, ENSG00000170606	HSPA4, ENSG00000135587	SMPD2, ENSG00000153896	ZNF599 
#> 	   ENSG00000080822	CLDND1, ENSG00000100083	GGA1, ENSG00000226742	HSBP1L1, ENSG00000213523	SRA1, ENSG00000166387	PPFIBP2, ENSG00000115128	SF3B14, ENSG00000108590	MED31, ENSG00000134851	TMEM165, ENSG00000164077	MON1A, ENSG00000167522	ANKRD11 
#> 	   ENSG00000072518	MARK2, ENSG00000251022	THAP9-AS1, ENSG00000095319	NUP188, ENSG00000262202	SNORD3D, ENSG00000257704	PRR24, ENSG00000176619	LMNB2, ENSG00000125637	PSD4, ENSG00000165934	CPSF2, ENSG00000132424	PNISR, ENSG00000171302	CANT1 
#> Computing nearest neighbor graph
#> Computing SNN
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 336
#> Number of edges: 23047
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.5000
#> Number of communities: 1
#> Elapsed time: 0 seconds
#> Normalizing layer: counts
#> Finding variable features for layer counts
#> Warning: pseudoinverse used at -2.1491
#> Warning: neighborhood radius 0.3166
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.031008
#> Centering and scaling data matrix
#> Warning: The following 1393 features requested have zero variance; running reduction without them: ENSG00000118292	C1orf54, ENSG00000134285	FKBP11, ENSG00000026025	VIM, ENSG00000165675	ENOX2, ENSG00000100983	GSS, ENSG00000133250	ZNF414, ENSG00000158481	CD1C, ENSG00000128272	ATF4, ENSG00000101220	C20orf27, ENSG00000108774	RAB5C, ENSG00000138382	METTL5, ENSG00000187257	RSBN1L, ENSG00000083844	ZNF264, ENSG00000030582	GRN, ENSG00000197816	CCDC180, ENSG00000153814	JAZF1, ENSG00000139278	GLIPR1, ENSG00000115233	PSMD14, ENSG00000167851	CD300A, ENSG00000095059	DHPS, ENSG00000115289	PCGF1, ENSG00000134516	DOCK2, ENSG00000135124	P2RX4, ENSG00000134202	GSTM3, ENSG00000115806	GORASP2, ENSG00000170876	TMEM43, ENSG00000173457	PPP1R14B, ENSG00000166145	SPINT1, ENSG00000134072	CAMK1, ENSG00000136147	PHF11, ENSG00000140691	ARMC5, ENSG00000243364	EFNA4, ENSG00000132716	DCAF8, ENSG00000197771	MCMBP, ENSG00000187118	CMC1, ENSG00000115649	CNPPD1, ENSG00000143797	MBOAT2, ENSG00000198453	ZNF568, ENSG00000170175	CHRNB1, ENSG00000120868	APAF1, ENSG00000214756	METTL12, ENSG00000116199	FAM20B, ENSG00000140090	SLC24A4, ENSG00000171793	CTPS1, ENSG00000006534	ALDH3B1, ENSG00000185291	IL3RA, ENSG00000266962	RP11-400F19.6, ENSG00000121931	LRIF1, ENSG00000148719	DNAJB12, ENSG00000171204	TMEM126B, ENSG00000141101	NOB1, ENSG00000226742	HSBP1L1, ENSG00000185658	BRWD1, ENSG00000104164	BLOC1S6, ENSG00000198700	IPO9, ENSG00000114648	KLHL18, ENSG00000204138	PHACTR4, ENSG00000158828	PINK1, ENSG00000119782	FKBP1B, ENSG00000131389	SLC6A6, ENSG00000173264	GPR137, ENSG00000110367	DDX6, ENSG00000184575	XPOT, ENSG00000140577	CRTC3, ENSG00000112799	LY86, ENSG00000066336	SPI1, ENSG00000121064	SCPEP1, ENSG00000178149	DALRD3, ENSG00000196262	PPIA, ENSG00000235999	RP11-403I13.8, ENSG00000088986	DYNLL1, ENSG00000168701	TMEM208, ENSG00000185507	IRF7, ENSG00000120742	SERP1, ENSG00000167642	SPINT2, ENSG00000155366	RHOC, ENSG00000104047	DTWD1, ENSG00000205155	PSENEN, ENSG00000168958	MFF, ENSG00000105755	ETHE1, ENSG00000115204	MPV17, ENSG00000102898	NUTF2, ENSG00000204252	HLA-DOA, ENSG00000239961	LILRA4, ENSG00000126759	CFP, ENSG00000173933	RBM4, ENSG00000133606	MKRN1, ENSG00000237441	RGL2, ENSG00000117335	CD46, ENSG00000167118	URM1, ENSG00000114857	NKTR, ENSG00000111348	ARHGDIB, ENSG00000166228	PCBD1, ENSG00000163131	CTSS, ENSG00000100348	TXN2, ENSG00000116906	GNPAT, ENSG00000141580	WDR45B, ENSG00000164327	RICTOR, ENSG00000108599	AKAP10, ENSG00000113282	CLINT1, ENSG00000088038	CNOT3, ENSG00000172757	CFL1, ENSG00000105058	FAM32A, ENSG00000106367	AP1S1, ENSG00000158714	SLAMF8, ENSG00000139725	RHOF, ENSG00000109919	MTCH2, ENSG00000139921	TMX1, ENSG00000179134	SAMD4B, ENSG00000271858	XXcos-LUCA11.4, ENSG00000251022	THAP9-AS1, ENSG00000111269	CREBL2, ENSG00000272686	RP11-390E23.6, ENSG00000127481	UBR4, ENSG00000198715	C1orf85, ENSG00000145687	SSBP2, ENSG00000055163	CYFIP2, ENSG00000114416	FXR1, ENSG00000167862	ICT1, ENSG00000092201	SUPT16H, ENSG00000184986	TMEM121, ENSG00000154473	BUB3, ENSG00000184674	GSTT1, ENSG00000084072	PPIE, ENSG00000177663	IL17RA, ENSG00000154930	ACSS1, ENSG00000197043	ANXA6, ENSG00000156411	C14orf2, ENSG00000166333	ILK, ENSG00000196642	RABL6, ENSG00000113712	CSNK1A1, ENSG00000186166	CCDC84, ENSG00000271888	RP11-560J1.2, ENSG00000258377	RP11-649E7.5, ENSG00000269089	AP003733.1, ENSG00000124209	RAB22A, ENSG00000241839	PLEKHO2, ENSG00000198130	HIBCH, ENSG00000257704	PRR24, ENSG00000166971	AKTIP, ENSG00000047932	GOPC, ENSG00000125637	PSD4, ENSG00000167522	ANKRD11, ENSG00000092036	HAUS4, ENSG00000145741	BTF3, ENSG00000128524	ATP6V1F, ENSG00000170921	TANC2, ENSG00000135392	DNAJC14, ENSG00000250510	GPR162, ENSG00000102174	PHEX, ENSG00000249348	UGDH-AS1, ENSG00000147475	ERLIN2, ENSG00000129521	EGLN3, ENSG00000107447	DNTT, ENSG00000268006	PTOV1-AS1, ENSG00000196505	GDAP2, ENSG00000168496	FEN1, ENSG00000260000	RP3-467N11.1, ENSG00000186522	SEPT10, ENSG00000100129	EIF3L, ENSG00000115053	NCL, ENSG00000102524	TNFSF13B, ENSG00000090554	FLT3LG, ENSG00000161526	SAP30BP, ENSG00000232442	CTD-3184A7.4, ENSG00000108591	DRG2, ENSG00000102309	PIN4, ENSG00000100811	YY1, ENSG00000166595	FAM96B, ENSG00000114631	PODXL2, ENSG00000175906	ARL4D, ENSG00000214367	HAUS3, ENSG00000162928	PEX13, ENSG00000197880	MDS2, ENSG00000259818	RP5-1024G6.7, ENSG00000125447	GGA3, ENSG00000165934	CPSF2, ENSG00000164031	DNAJB14, ENSG00000259375	RP11-815J21.2, ENSG00000182685	BRICD5, ENSG00000124334	IL9R, ENSG00000204954	C12orf73, ENSG00000155324	GRAMD3, ENSG00000214293	RSBN1L-AS1, ENSG00000124813	RUNX2, ENSG00000229084	LINC00226, ENSG00000159556	ISL2, ENSG00000167196	FBXO22, ENSG00000115520	COQ10B, ENSG00000100068	LRP5L, ENSG00000162959	MEMO1, ENSG00000141424	SLC39A6, ENSG00000261105	RP11-173B14.5, ENSG00000247121	CTD-2260A17.2, ENSG00000104524	PYCRL, ENSG00000254802	AC022182.3, ENSG00000160216	AGPAT3, ENSG00000173825	TIGD3, ENSG00000155827	RNF20, ENSG00000157456	CCNB2, ENSG00000100399	CHADL, ENSG00000114923	SLC4A3, ENSG00000146143	PRIM2, ENSG00000135913	USP37, ENSG00000106683	LIMK1, ENSG00000227403	AC009299.3, ENSG00000186812	ZNF397, ENSG00000140525	FANCI, ENSG00000065029	ZNF76, ENSG00000164332	UBLCP1, ENSG00000110048	OSBP, ENSG00000213967	ZNF726, ENSG00000267475	CTD-2538C1.2, ENSG00000113812	ACTR8, ENSG00000165983	PTER, ENSG00000136811	ODF2, ENSG00000114268	PFKFB4, ENSG00000113391	FAM172A, ENSG00000230724	LINC01001, ENSG00000123908	AGO2, ENSG00000135093	USP30, ENSG00000147874	HAUS6, ENSG00000130787	HIP1R, ENSG00000260077	RP11-254F7.2, ENSG00000248905	FMN1, ENSG00000204388	HSPA1B, ENSG00000183018	SPNS2, ENSG00000181894	ZNF329, ENSG00000177335	C8orf31, ENSG00000205517	RGL3, ENSG00000112039	FANCE, ENSG00000166548	TK2, ENSG00000132010	ZNF20, ENSG00000183647	ZNF530, ENSG00000171448	ZBTB26, ENSG00000247199	RP11-373N22.3, ENSG00000087053	MTMR2, ENSG00000105088	OLFM2, ENSG00000106524	ANKMY2, ENSG00000197183	C20orf112, ENSG00000132549	VPS13B, ENSG00000137656	BUD13, ENSG00000128191	DGCR8, ENSG00000215615	AL354822.1, ENSG00000174307	PHLDA3, ENSG00000119227	PIGZ, ENSG00000137804	NUSAP1, ENSG00000178691	SUZ12, ENSG00000072518	MARK2, ENSG00000004777	ARHGAP33, ENSG00000011258	MBTD1, ENSG00000272986	RP11-46J23.1, ENSG00000143340	FAM163A, ENSG00000227908	CTD-2031P19.3, ENSG00000269996	RP11-343N15.5, ENSG00000166387	PPFIBP2, ENSG00000169905	TOR1AIP2, ENSG00000228265	RP5-1125A11.1, ENSG00000111224	PARP11, ENSG00000231789	RP11-558F24.4, ENSG00000013288	MAN2B2, ENSG00000229729	RP11-159G9.5, ENSG00000152127	MGAT5, ENSG00000169981	ZNF35, ENSG00000126368	NR1D1, ENSG00000122482	ZNF644, ENSG00000145414	NAF1, ENSG00000071655	MBD3, ENSG00000166439	RNF169, ENSG00000009830	POMT2, ENSG00000234230	ZFX-AS1, ENSG00000137720	C11orf1, ENSG00000152580	IGSF10, ENSG00000076685	NT5C2, ENSG00000198105	ZNF248, ENSG00000246898	LINC00920, ENSG00000116574	RHOU, ENSG00000133392	MYH11, ENSG00000214900	C14orf182, ENSG00000156599	ZDHHC5, ENSG00000122481	RWDD3, ENSG00000154359	LONRF1, ENSG00000145416	MARCH1, ENSG00000196365	LONP1, ENSG00000168228	ZCCHC4, ENSG00000159899	NPR2, ENSG00000261505	LA16c-358B7.3, ENSG00000231435	AC011747.3, ENSG00000172831	CES2, ENSG00000158793	NIT1, ENSG00000157106	SMG1, ENSG00000178971	CTC1, ENSG00000163486	SRGAP2, ENSG00000273139	XXbac-B444P24.14, ENSG00000175866	BAIAP2, ENSG00000170542	SERPINB9, ENSG00000188917	TRMT2B, ENSG00000230074	RP11-195F19.9, ENSG00000123815	ADCK4, ENSG00000043514	TRIT1, ENSG00000164077	MON1A, ENSG00000228434	AC004951.6, ENSG00000080200	CRYBG3, ENSG00000261687	RP11-473C18.3, ENSG00000184007	PTP4A2, ENSG00000099337	KCNK6, ENSG00000267519	MIR24-2, ENSG00000140307	GTF2A2, ENSG00000196230	TUBB, ENSG00000105993	DNAJB6, ENSG00000173992	CCS, ENSG00000105341	ATP5SL, ENSG00000167578	RAB4B, ENSG00000148335	NTMT1, ENSG00000181396	OGFOD3, ENSG00000103335	PIEZO1, ENSG00000173327	MAP3K11, ENSG00000146729	GBAS, ENSG00000118900	UBN1, ENSG00000198189	HSD17B11, ENSG00000241837	ATP5O, ENSG00000167208	SNX20, ENSG00000109062	SLC9A3R1, ENSG00000198355	PIM3, ENSG00000039650	PNKP, ENSG00000123240	OPTN, ENSG00000108826	MRPL27, ENSG00000187514	PTMA, ENSG00000169032	MAP2K1, ENSG00000183431	SF3A3, ENSG00000166326	TRIM44, ENSG00000069998	CECR5, ENSG00000080189	SLC35C2, ENSG00000117899	MESDC2, ENSG00000187266	EPOR, ENSG00000224051	GLTPD1, ENSG000001
#> PC_ 1 
#> Positive:  ENSG00000198821	CD247, ENSG00000144746	ARL6IP5, ENSG00000180096	SEPT1, ENSG00000027075	PRKCH, ENSG00000088766	CRLS1, ENSG00000132382	MYBBP1A, ENSG00000136717	BIN1, ENSG00000256060	TRAPPC2P1, ENSG00000161204	ABCF3, ENSG00000260083	MIR4519 
#> 	   ENSG00000176248	ANAPC2, ENSG00000099992	TBC1D10A, ENSG00000204564	C6orf136, ENSG00000235106	LINC00094, ENSG00000116221	MRPL37, ENSG00000189339	SLC35E2B, ENSG00000121897	LIAS, ENSG00000157538	DSCR3, ENSG00000143570	SLC39A1, ENSG00000186265	BTLA 
#> 	   ENSG00000263508	RP11-963H4.3, ENSG00000167483	FAM129C, ENSG00000177398	UMODL1, ENSG00000175893	ZDHHC21, ENSG00000048991	R3HDM1, ENSG00000174059	CD34, ENSG00000184154	LRTOMT, ENSG00000153107	ANAPC1, ENSG00000143167	GPA33, ENSG00000142892	PIGK 
#> Negative:  ENSG00000223865	HLA-DPB1, ENSG00000182359	KBTBD3, ENSG00000156170	NDUFAF6, ENSG00000171848	RRM2, ENSG00000164305	CASP3, ENSG00000119411	BSPRY, ENSG00000197780	TAF13, ENSG00000085788	DDHD2, ENSG00000075234	TTC38, ENSG00000173064	HECTD4 
#> 	   ENSG00000025708	TYMP, ENSG00000204482	LST1, ENSG00000124570	SERPINB6, ENSG00000159884	CCDC107, ENSG00000213676	ATF6B, ENSG00000198951	NAGA, ENSG00000099956	SMARCB1, ENSG00000136152	COG3, ENSG00000137074	APTX, ENSG00000104774	MAN2B1 
#> 	   ENSG00000184451	CCR10, ENSG00000215021	PHB2, ENSG00000130475	FCHO1, ENSG00000229117	RPL41, ENSG00000129534	MIS18BP1, ENSG00000137100	DCTN3, ENSG00000100908	EMC9, ENSG00000167536	DHRS13, ENSG00000272269	RP11-500C11.3, ENSG00000163040	CCDC74A 
#> PC_ 2 
#> Positive:  ENSG00000223865	HLA-DPB1, ENSG00000124570	SERPINB6, ENSG00000204482	LST1, ENSG00000124357	NAGK, ENSG00000197223	C1D, ENSG00000100908	EMC9, ENSG00000025708	TYMP, ENSG00000173064	HECTD4, ENSG00000130475	FCHO1, ENSG00000161243	FBXO27 
#> 	   ENSG00000168303	MPLKIP, ENSG00000198951	NAGA, ENSG00000163251	FZD5, ENSG00000139719	VPS33A, ENSG00000273080	RP11-301O19.1, ENSG00000160326	SLC2A6, ENSG00000174652	ZNF266, ENSG00000101265	RASSF2, ENSG00000136527	TRA2B, ENSG00000238227	C9orf69 
#> 	   ENSG00000087076	HSD17B14, ENSG00000247240	UBL7-AS1, ENSG00000141452	C18orf8, ENSG00000090621	PABPC4, ENSG00000198498	TMA16, ENSG00000084234	APLP2, ENSG00000155330	C16orf87, ENSG00000104885	DOT1L, ENSG00000159314	ARHGAP27, ENSG00000145545	SRD5A1 
#> Negative:  ENSG00000085788	DDHD2, ENSG00000119411	BSPRY, ENSG00000197780	TAF13, ENSG00000171848	RRM2, ENSG00000156170	NDUFAF6, ENSG00000164305	CASP3, ENSG00000184451	CCR10, ENSG00000075234	TTC38, ENSG00000100124	ANKRD54, ENSG00000180096	SEPT1 
#> 	   ENSG00000143353	LYPLAL1, ENSG00000243279	PRAF2, ENSG00000080822	CLDND1, ENSG00000182359	KBTBD3, ENSG00000136152	COG3, ENSG00000241370	RPP21, ENSG00000126756	UXT, ENSG00000139233	LLPH, ENSG00000254470	AP5B1, ENSG00000198821	CD247 
#> 	   ENSG00000213676	ATF6B, ENSG00000099330	OCEL1, ENSG00000079246	XRCC5, ENSG00000135521	LTV1, ENSG00000166260	COX11, ENSG00000137074	APTX, ENSG00000104774	MAN2B1, ENSG00000132424	PNISR, ENSG00000027075	PRKCH, ENSG00000184752	NDUFA12 
#> PC_ 3 
#> Positive:  ENSG00000177398	UMODL1, ENSG00000263508	RP11-963H4.3, ENSG00000167483	FAM129C, ENSG00000048991	R3HDM1, ENSG00000174059	CD34, ENSG00000175893	ZDHHC21, ENSG00000138138	ATAD1, ENSG00000256060	TRAPPC2P1, ENSG00000143570	SLC39A1, ENSG00000155252	PI4K2A 
#> 	   ENSG00000109787	KLF3, ENSG00000135916	ITM2C, ENSG00000099991	CABIN1, ENSG00000249222	ATP5L2, ENSG00000118705	RPN2, ENSG00000180096	SEPT1, ENSG00000243279	PRAF2, ENSG00000105619	TFPT, ENSG00000136527	TRA2B, ENSG00000185650	ZFP36L1 
#> 	   ENSG00000116030	SUMO1, ENSG00000170540	ARL6IP1, ENSG00000168090	COPS6, ENSG00000144746	ARL6IP5, ENSG00000247240	UBL7-AS1, ENSG00000087076	HSD17B14, ENSG00000238227	C9orf69, ENSG00000118680	MYL12B, ENSG00000161243	FBXO27, ENSG00000197223	C1D 
#> Negative:  ENSG00000025708	TYMP, ENSG00000229117	RPL41, ENSG00000173221	GLRX, ENSG00000205246	RPSAP58, ENSG00000084234	APLP2, ENSG00000127774	EMC6, ENSG00000136867	SLC31A2, ENSG00000164733	CTSB, ENSG00000124357	NAGK, ENSG00000126749	EMG1 
#> 	   ENSG00000170606	HSPA4, ENSG00000131378	RFTN1, ENSG00000105193	RPS16, ENSG00000180509	KCNE1, ENSG00000137474	MYO7A, ENSG00000149532	CPSF7, ENSG00000117505	DR1, ENSG00000146828	SLC12A9, ENSG00000169612	FAM103A1, ENSG00000091009	RBM27 
#> 	   ENSG00000160888	IER2, ENSG00000115828	QPCT, ENSG00000105443	CYTH2, ENSG00000184277	TM2D3, ENSG00000260651	AF213884.2, ENSG00000165406	MARCH8, ENSG00000181982	CCDC149, ENSG00000138646	HERC5, ENSG00000091106	NLRC4, ENSG00000160796	NBEAL2 
#> PC_ 4 
#> Positive:  ENSG00000136717	BIN1, ENSG00000126756	UXT, ENSG00000164088	PPM1M, ENSG00000170248	PDCD6IP, ENSG00000174652	ZNF266, ENSG00000118705	RPN2, ENSG00000143575	HAX1, ENSG00000162244	RPL29, ENSG00000074755	ZZEF1, ENSG00000198821	CD247 
#> 	   ENSG00000161243	FBXO27, ENSG00000105829	BET1, ENSG00000260796	RP11-1348G14.5, ENSG00000134882	UBAC2, ENSG00000173064	HECTD4, ENSG00000155330	C16orf87, ENSG00000142892	PIGK, ENSG00000116030	SUMO1, ENSG00000197223	C1D, ENSG00000172717	FAM71D 
#> 	   ENSG00000238227	C9orf69, ENSG00000247240	UBL7-AS1, ENSG00000087076	HSD17B14, ENSG00000107771	CCSER2, ENSG00000180096	SEPT1, ENSG00000132424	PNISR, ENSG00000197157	SND1, ENSG00000080822	CLDND1, ENSG00000153130	SCOC, ENSG00000198498	TMA16 
#> Negative:  ENSG00000204482	LST1, ENSG00000025708	TYMP, ENSG00000223865	HLA-DPB1, ENSG00000136867	SLC31A2, ENSG00000155252	PI4K2A, ENSG00000229117	RPL41, ENSG00000164733	CTSB, ENSG00000065989	PDE4A, ENSG00000180509	KCNE1, ENSG00000174059	CD34 
#> 	   ENSG00000175893	ZDHHC21, ENSG00000263508	RP11-963H4.3, ENSG00000048991	R3HDM1, ENSG00000177398	UMODL1, ENSG00000167483	FAM129C, ENSG00000170606	HSPA4, ENSG00000127774	EMC6, ENSG00000184277	TM2D3, ENSG00000124357	NAGK, ENSG00000115828	QPCT 
#> 	   ENSG00000127616	SMARCA4, ENSG00000138646	HERC5, ENSG00000091106	NLRC4, ENSG00000143570	SLC39A1, ENSG00000091009	RBM27, ENSG00000249222	ATP5L2, ENSG00000137474	MYO7A, ENSG00000149532	CPSF7, ENSG00000136152	COG3, ENSG00000075234	TTC38 
#> PC_ 5 
#> Positive:  ENSG00000074755	ZZEF1, ENSG00000257303	RP11-977G19.11, ENSG00000096968	JAK2, ENSG00000063761	ADCK1, ENSG00000197471	SPN, ENSG00000107771	CCSER2, ENSG00000107581	EIF3A, ENSG00000143207	RFWD2, ENSG00000121644	DESI2, ENSG00000260796	RP11-1348G14.5 
#> 	   ENSG00000105829	BET1, ENSG00000198951	NAGA, ENSG00000167766	ZNF83, ENSG00000109534	GAR1, ENSG00000129158	SERGEF, ENSG00000073803	MAP3K13, ENSG00000099991	CABIN1, ENSG00000105619	TFPT, ENSG00000164088	PPM1M, ENSG00000172717	FAM71D 
#> 	   ENSG00000134882	UBAC2, ENSG00000159884	CCDC107, ENSG00000170248	PDCD6IP, ENSG00000129534	MIS18BP1, ENSG00000099956	SMARCB1, ENSG00000204482	LST1, ENSG00000116221	MRPL37, ENSG00000158805	ZNF276, ENSG00000142892	PIGK, ENSG00000223865	HLA-DPB1 
#> Negative:  ENSG00000197223	C1D, ENSG00000139719	VPS33A, ENSG00000100908	EMC9, ENSG00000115128	SF3B14, ENSG00000163251	FZD5, ENSG00000168303	MPLKIP, ENSG00000145545	SRD5A1, ENSG00000161243	FBXO27, ENSG00000273080	RP11-301O19.1, ENSG00000104885	DOT1L 
#> 	   ENSG00000238227	C9orf69, ENSG00000087076	HSD17B14, ENSG00000247240	UBL7-AS1, ENSG00000130475	FCHO1, ENSG00000173064	HECTD4, ENSG00000143575	HAX1, ENSG00000169710	FASN, ENSG00000141452	C18orf8, ENSG00000090621	PABPC4, ENSG00000197912	SPG7 
#> 	   ENSG00000102967	DHODH, ENSG00000101265	RASSF2, ENSG00000249222	ATP5L2, ENSG00000153130	SCOC, ENSG00000241990	RP6-109B7.3, ENSG00000159479	MED8, ENSG00000213676	ATF6B, ENSG00000172366	FAM195A, ENSG00000160326	SLC2A6, ENSG00000135587	SMPD2 
#> Computing nearest neighbor graph
#> Computing SNN
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 136
#> Number of edges: 6348
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.5848
#> Number of communities: 2
#> Elapsed time: 0 seconds
#> Normalizing layer: counts
#> Finding variable features for layer counts
#> Warning: pseudoinverse used at -2.5336
#> Warning: neighborhood radius 0.49481
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 0.090619
#> Centering and scaling data matrix
#> Warning: The following 1304 features requested have zero variance; running reduction without them: ENSG00000164081	TEX264, ENSG00000243279	PRAF2, ENSG00000142784	WDTC1, ENSG00000168496	FEN1, ENSG00000187118	CMC1, ENSG00000102575	ACP5, ENSG00000165671	NSD1, ENSG00000111666	CHPT1, ENSG00000262202	SNORD3D, ENSG00000235106	LINC00094, ENSG00000198700	IPO9, ENSG00000160741	CRTC2, ENSG00000120662	MTRF1, ENSG00000131759	RARA, ENSG00000257698	RP11-620J15.3, ENSG00000144824	PHLDB2, ENSG00000189134	NKAPL, ENSG00000137404	NRM, ENSG00000130684	ZNF337, ENSG00000113812	ACTR8, ENSG00000101639	CEP192, ENSG00000157557	ETS2, ENSG00000090447	TFAP4, ENSG00000117643	MAN1C1, ENSG00000225439	BOLA3-AS1, ENSG00000039560	RAI14, ENSG00000188001	TPRG1, ENSG00000213762	ZNF134, ENSG00000178971	CTC1, ENSG00000160888	IER2, ENSG00000185344	ATP6V0A2, ENSG00000100811	YY1, ENSG00000169249	ZRSR2, ENSG00000100554	ATP6V1D, ENSG00000120833	SOCS2, ENSG00000164306	PRIMPOL, ENSG00000180209	MYLPF, ENSG00000166548	TK2, ENSG00000004487	KDM1A, ENSG00000198453	ZNF568, ENSG00000115042	FAHD2A, ENSG00000152582	SPEF2, ENSG00000131389	SLC6A6, ENSG00000170919	TPT1-AS1, ENSG00000156599	ZDHHC5, ENSG00000125637	PSD4, ENSG00000166145	SPINT1, ENSG00000109062	SLC9A3R1, ENSG00000198546	ZNF511, ENSG00000173992	CCS, ENSG00000198821	CD247, ENSG00000214756	METTL12, ENSG00000132716	DCAF8, ENSG00000140545	MFGE8, ENSG00000120868	APAF1, ENSG00000215039	CD27-AS1, ENSG00000100083	GGA1, ENSG00000025708	TYMP, ENSG00000155330	C16orf87, ENSG00000197771	MCMBP, ENSG00000140987	ZSCAN32, ENSG00000153071	DAB2, ENSG00000176274	SLC25A53, ENSG00000180957	PITPNB, ENSG00000155366	RHOC, ENSG00000129534	MIS18BP1, ENSG00000137074	APTX, ENSG00000168818	STX18, ENSG00000164077	MON1A, ENSG00000099246	RAB18, ENSG00000167851	CD300A, ENSG00000135521	LTV1, ENSG00000105373	GLTSCR2, ENSG00000185347	C14orf80, ENSG00000121064	SCPEP1, ENSG00000131013	PPIL4, ENSG00000105619	TFPT, ENSG00000139921	TMX1, ENSG00000176593	CTD-2368P22.1, ENSG00000154930	ACSS1, ENSG00000137193	PIM1, ENSG00000110367	DDX6, ENSG00000106355	LSM5, ENSG00000155827	RNF20, ENSG00000100949	RABGGTA, ENSG00000141873	SLC39A3, ENSG00000163131	CTSS, ENSG00000058272	PPP1R12A, ENSG00000156973	PDE6D, ENSG00000124357	NAGK, ENSG00000136940	PDCL, ENSG00000232442	CTD-3184A7.4, ENSG00000160087	UBE2J2, ENSG00000139233	LLPH, ENSG00000167118	URM1, ENSG00000163728	TTC14, ENSG00000256060	TRAPPC2P1, ENSG00000068745	IP6K2, ENSG00000198860	TSEN15, ENSG00000084234	APLP2, ENSG00000112062	MAPK14, ENSG00000100564	PIGH, ENSG00000196262	PPIA, ENSG00000198189	HSD17B11, ENSG00000197471	SPN, ENSG00000170759	KIF5B, ENSG00000102898	NUTF2, ENSG00000129158	SERGEF, ENSG00000148335	NTMT1, ENSG00000121749	TBC1D15, ENSG00000170876	TMEM43, ENSG00000099917	MED15, ENSG00000148950	IMMP1L, ENSG00000141101	NOB1, ENSG00000198574	SH2D1B, ENSG00000197912	SPG7, ENSG00000121931	LRIF1, ENSG00000073614	KDM5A, ENSG00000156411	C14orf2, ENSG00000113712	CSNK1A1, ENSG00000166260	COX11, ENSG00000127838	PNKD, ENSG00000167543	TP53I13, ENSG00000174744	BRMS1, ENSG00000153130	SCOC, ENSG00000167862	ICT1, ENSG00000251022	THAP9-AS1, ENSG00000139278	GLIPR1, ENSG00000143553	SNAPIN, ENSG00000143815	LBR, ENSG00000055163	CYFIP2, ENSG00000243224	RP5-1157M23.2, ENSG00000118292	C1orf54, ENSG00000198517	MAFK, ENSG00000140675	SLC5A2, ENSG00000136717	BIN1, ENSG00000184752	NDUFA12, ENSG00000157538	DSCR3, ENSG00000108826	MRPL27, ENSG00000162642	C1orf52, ENSG00000120948	TARDBP, ENSG00000105193	RPS16, ENSG00000115204	MPV17, ENSG00000187266	EPOR, ENSG00000114631	PODXL2, ENSG00000214960	ISPD, ENSG00000099364	FBXL19, ENSG00000175322	ZNF519, ENSG00000242767	ZBTB20-AS4, ENSG00000179134	SAMD4B, ENSG00000259818	RP5-1024G6.7, ENSG00000117479	SLC19A2, ENSG00000272663	RP11-191L17.1, ENSG00000123243	ITIH5, ENSG00000152969	JAKMIP1, ENSG00000124334	IL9R, ENSG00000214293	RSBN1L-AS1, ENSG00000160326	SLC2A6, ENSG00000053524	MCF2L2, ENSG00000175787	ZNF169, ENSG00000145908	ZNF300, ENSG00000198551	ZNF627, ENSG00000136867	SLC31A2, ENSG00000197182	FLJ27365, ENSG00000186281	GPAT2, ENSG00000112029	FBXO5, ENSG00000099337	KCNK6, ENSG00000196357	ZNF565, ENSG00000134460	IL2RA, ENSG00000130714	POMT1, ENSG00000205106	CTD-2210P24.4, ENSG00000146143	PRIM2, ENSG00000272841	RP3-428L16.2, ENSG00000186812	ZNF397, ENSG00000251637	RP11-119D9.1, ENSG00000163291	PAQR3, ENSG00000149260	CAPN5, ENSG00000162520	SYNC, ENSG00000257497	RP11-585P4.5, ENSG00000267002	RP11-242D8.1, ENSG00000167513	CDT1, ENSG00000184270	HIST2H2AB, ENSG00000147166	ITGB1BP2, ENSG00000245614	DDX11-AS1, ENSG00000167483	FAM129C, ENSG00000178773	CPNE7, ENSG00000171848	RRM2, ENSG00000248905	FMN1, ENSG00000126821	SGPP1, ENSG00000133466	C1QTNF6, ENSG00000180089	TMEM86B, ENSG00000116199	FAM20B, ENSG00000140750	ARHGAP17, ENSG00000165684	SNAPC4, ENSG00000227345	PARG, ENSG00000271856	RP11-861A13.4, ENSG00000215421	ZNF407, ENSG00000152253	SPC25, ENSG00000163960	UBXN7, ENSG00000213380	COG8, ENSG00000214300	SPDYE3, ENSG00000110723	EXPH5, ENSG00000168754	FAM178B, ENSG00000160113	NR2F6, ENSG00000181751	C5orf30, ENSG00000198920	KIAA0753, ENSG00000021574	SPAST, ENSG00000040608	RTN4R, ENSG00000006534	ALDH3B1, ENSG00000128266	GNAZ, ENSG00000125726	CD70, ENSG00000184349	EFNA5, ENSG00000143412	ANXA9, ENSG00000162616	DNAJB4, ENSG00000197780	TAF13, ENSG00000169981	ZNF35, ENSG00000132470	ITGB4, ENSG00000173559	NABP1, ENSG00000175155	YPEL2, ENSG00000272382	CTD-2035E11.4, ENSG00000179397	C1orf101, ENSG00000253704	RP11-267M23.4, ENSG00000174939	ASPHD1, ENSG00000051382	PIK3CB, ENSG00000070047	PHRF1, ENSG00000039523	FAM65A, ENSG00000125885	MCM8, ENSG00000147454	SLC25A37, ENSG00000197816	CCDC180, ENSG00000160588	MPZL3, ENSG00000261737	RP4-612B15.3, ENSG00000231177	LINC00852, ENSG00000156030	ELMSAN1, ENSG00000092036	HAUS4, ENSG00000105610	KLF1, ENSG00000182636	NDN, ENSG00000173276	ZBTB21, ENSG00000189419	SPATA41, ENSG00000177663	IL17RA, ENSG00000159899	NPR2, ENSG00000163328	GPR155, ENSG00000261505	LA16c-358B7.3, ENSG00000258122	RP11-61A14.1, ENSG00000167608	TMC4, ENSG00000184886	PIGW, ENSG00000158079	PTPDC1, ENSG00000164087	POC1A, ENSG00000259349	RP11-15E18.1, ENSG00000233806	AC131097.3, ENSG00000247796	CTD-2366F13.1, ENSG00000228434	AC004951.6, ENSG00000270189	RP11-258C19.7, ENSG00000182534	MXRA7, ENSG00000063761	ADCK1, ENSG00000134884	ARGLU1, ENSG00000143575	HAX1, ENSG00000172757	CFL1, ENSG00000141522	ARHGDIA, ENSG00000088986	DYNLL1, ENSG00000080189	SLC35C2, ENSG00000159479	MED8, ENSG00000235999	RP11-403I13.8, ENSG00000160688	FLAD1, ENSG00000136527	TRA2B, ENSG00000185650	ZFP36L1, ENSG00000075234	TTC38, ENSG00000108774	RAB5C, ENSG00000173221	GLRX, ENSG00000102409	BEX4, ENSG00000105058	FAM32A, ENSG00000132424	PNISR, ENSG00000107771	CCSER2, ENSG00000107581	EIF3A, ENSG00000154473	BUB3, ENSG00000149311	ATM, ENSG00000108010	GLRX3, ENSG00000132382	MYBBP1A, ENSG00000185305	ARL15, ENSG00000197535	MYO5A, ENSG00000156313	RPGR, ENSG00000165996	PTPLA, ENSG00000124813	RUNX2, ENSG00000170921	TANC2, ENSG00000153767	GTF2E1, ENSG00000154734	ADAMTS1, ENSG00000135913	USP37, ENSG00000270195	RP11-572O17.1, ENSG00000119986	AVPI1, ENSG00000106608	URGCP, ENSG00000156500	FAM122C, ENSG00000213967	ZNF726, ENSG00000165359	DDX26B, ENSG00000253210	RP11-809O17.1, ENSG00000147119	CHST7, ENSG00000261079	RP11-252A24.3, ENSG00000270964	RP11-502I4.3, ENSG00000186166	CCDC84, ENSG00000143363	PRUNE, ENSG00000125247	TMTC4, ENSG00000132010	ZNF20, ENSG00000183647	ZNF530, ENSG00000225697	SLC26A6, ENSG00000161091	MFSD12, ENSG00000116273	PHF13, ENSG00000174093	RP11-1407O15.2, ENSG00000004777	ARHGAP33, ENSG00000156876	SASS6, ENSG00000243364	EFNA4, ENSG00000187626	ZKSCAN4, ENSG00000106003	LFNG, ENSG00000160285	LSS, ENSG00000266964	FXYD1, ENSG00000171793	CTPS1, ENSG00000111252	SH2B3, ENSG00000154764	WNT7A, ENSG00000135966	TGFBRAP1, ENSG00000009830	POMT2, ENSG00000143458	GABPB2, ENSG00000198105	ZNF248, ENSG00000117000	RLF, ENSG00000266962	RP11-400F19.6, ENSG00000179627	ZBTB42, ENSG00000173064	HECTD4, ENSG00000156162	DPY19L4, ENSG00000148516	ZEB1, ENSG00000170260	ZNF212, ENSG00000137802	MAPKBP1, ENSG00000162620	LRRIQ3, ENSG00000020256	ZFP64, ENSG00000243819	RN7SL832P, ENSG00000184481	FOXO4, ENSG00000166947	EPB42, ENSG00000124459	ZNF45, ENSG00000136436	CALCOCO2, ENSG00000105755	ETHE1, ENSG00000180096	SE
#> PC_ 1 
#> Positive:  ENSG00000101474	APMAP, ENSG00000213741	RPS29, ENSG00000158805	ZNF276, ENSG00000184007	PTP4A2, ENSG00000130770	ATPIF1, ENSG00000186010	NDUFA13, ENSG00000139725	RHOF, ENSG00000118680	MYL12B, ENSG00000120699	EXOSC8, ENSG00000124570	SERPINB6 
#> 	   ENSG00000130703	OSBPL2, ENSG00000099956	SMARCB1, ENSG00000099330	OCEL1, ENSG00000178691	SUZ12, ENSG00000168701	TMEM208, ENSG00000125995	ROMO1, ENSG00000241370	RPP21, ENSG00000141002	TCF25, ENSG00000102309	PIN4, ENSG00000123240	OPTN 
#> 	   ENSG00000113391	FAM172A, ENSG00000164031	DNAJB14, ENSG00000164305	CASP3, ENSG00000020633	RUNX3, ENSG00000116237	ICMT, ENSG00000075089	ACTR6, ENSG00000181847	TIGIT, ENSG00000170540	ARL6IP1, ENSG00000174917	C19orf70, ENSG00000100348	TXN2 
#> Negative:  ENSG00000139193	CD27, ENSG00000162244	RPL29, ENSG00000026025	VIM, ENSG00000090554	FLT3LG, ENSG00000063177	RPL18, ENSG00000095059	DHPS, ENSG00000153391	INO80C, ENSG00000100129	EIF3L, ENSG00000134202	GSTM3, ENSG00000173327	MAP3K11 
#> 	   ENSG00000197880	MDS2, ENSG00000204482	LST1, ENSG00000198355	PIM3, ENSG00000168101	NUDT16L1, ENSG00000205246	RPSAP58, ENSG00000179598	PLD6, ENSG00000266947	RP11-799D4.4, ENSG00000253276	CCDC71L, ENSG00000131368	MRPS25, ENSG00000167536	DHRS13 
#> 	   ENSG00000180376	CCDC66, ENSG00000105088	OLFM2, ENSG00000186376	ZNF75D, ENSG00000254802	AC022182.3, ENSG00000111877	MCM9, ENSG00000168209	DDIT4, ENSG00000169100	SLC25A6, ENSG00000163162	RNF149, ENSG00000146828	SLC12A9, ENSG00000137411	VARS2 
#> PC_ 2 
#> Positive:  ENSG00000153107	ANAPC1, ENSG00000136811	ODF2, ENSG00000113391	FAM172A, ENSG00000197841	ZNF181, ENSG00000223759	RP11-353N4.4, ENSG00000169499	PLEKHA2, ENSG00000273117	AC144652.1, ENSG00000100359	SGSM3, ENSG00000211456	SACM1L, ENSG00000064651	SLC12A2 
#> 	   ENSG00000241839	PLEKHO2, ENSG00000125482	TTF1, ENSG00000153140	CETN3, ENSG00000072364	AFF4, ENSG00000106524	ANKMY2, ENSG00000240291	RP11-499P20.2, ENSG00000135365	PHF21A, ENSG00000142892	PIGK, ENSG00000259230	CTD-2555C10.3, ENSG00000115128	SF3B14 
#> 	   ENSG00000100983	GSS, ENSG00000075089	ACTR6, ENSG00000105993	DNAJB6, ENSG00000184277	TM2D3, ENSG00000164305	CASP3, ENSG00000148218	ALAD, ENSG00000125995	ROMO1, ENSG00000130770	ATPIF1, ENSG00000105329	TGFB1, ENSG00000126749	EMG1 
#> Negative:  ENSG00000130703	OSBPL2, ENSG00000206053	HN1L, ENSG00000186376	ZNF75D, ENSG00000116237	ICMT, ENSG00000170248	PDCD6IP, ENSG00000099991	CABIN1, ENSG00000164031	DNAJB14, ENSG00000115520	COQ10B, ENSG00000143353	LYPLAL1, ENSG00000166228	PCBD1 
#> 	   ENSG00000120742	SERP1, ENSG00000084072	PPIE, ENSG00000082805	ERC1, ENSG00000102309	PIN4, ENSG00000054148	PHPT1, ENSG00000105866	SP4, ENSG00000060339	CCAR1, ENSG00000171204	TMEM126B, ENSG00000030582	GRN, ENSG00000146828	SLC12A9 
#> 	   ENSG00000164088	PPM1M, ENSG00000272269	RP11-500C11.3, ENSG00000114416	FXR1, ENSG00000106603	COA1, ENSG00000175756	AURKAIP1, ENSG00000137601	NEK1, ENSG00000165675	ENOX2, ENSG00000102900	NUP93, ENSG00000155592	ZKSCAN2, ENSG00000181847	TIGIT 
#> PC_ 3 
#> Positive:  ENSG00000082701	GSK3B, ENSG00000126749	EMG1, ENSG00000116497	S100PBP, ENSG00000137411	VARS2, ENSG00000170540	ARL6IP1, ENSG00000213741	RPS29, ENSG00000150347	ARID5B, ENSG00000101052	IFT52, ENSG00000102309	PIN4, ENSG00000237441	RGL2 
#> 	   ENSG00000134014	ELP3, ENSG00000100983	GSS, ENSG00000169032	MAP2K1, ENSG00000105329	TGFB1, ENSG00000104518	GSDMD, ENSG00000270170	NCBP2-AS2, ENSG00000226742	HSBP1L1, ENSG00000131368	MRPS25, ENSG00000228857	AC104653.1, ENSG00000153814	JAZF1 
#> 	   ENSG00000214106	PAXIP1-AS2, ENSG00000060339	CCAR1, ENSG00000164733	CTSB, ENSG00000172500	FIBP, ENSG00000187514	PTMA, ENSG00000125482	TTF1, ENSG00000168090	COPS6, ENSG00000197948	FCHSD1, ENSG00000129317	PUS7L, ENSG00000167522	ANKRD11 
#> Negative:  ENSG00000153107	ANAPC1, ENSG00000113391	FAM172A, ENSG00000197841	ZNF181, ENSG00000026025	VIM, ENSG00000106524	ANKMY2, ENSG00000273117	AC144652.1, ENSG00000169100	SLC25A6, ENSG00000240291	RP11-499P20.2, ENSG00000063177	RPL18, ENSG00000223759	RP11-353N4.4 
#> 	   ENSG00000072364	AFF4, ENSG00000156170	NDUFAF6, ENSG00000174100	MRPL45, ENSG00000130703	OSBPL2, ENSG00000100359	SGSM3, ENSG00000135365	PHF21A, ENSG00000100348	TXN2, ENSG00000168209	DDIT4, ENSG00000137601	NEK1, ENSG00000272831	RP11-792A8.4 
#> 	   ENSG00000161526	SAP30BP, ENSG00000206053	HN1L, ENSG00000162244	RPL29, ENSG00000099330	OCEL1, ENSG00000146828	SLC12A9, ENSG00000139193	CD27, ENSG00000120699	EXOSC8, ENSG00000184277	TM2D3, ENSG00000171204	TMEM126B, ENSG00000116237	ICMT 
#> PC_ 4 
#> Positive:  ENSG00000196591	HDAC2, ENSG00000168090	COPS6, ENSG00000118518	RNF146, ENSG00000105993	DNAJB6, ENSG00000099991	CABIN1, ENSG00000240291	RP11-499P20.2, ENSG00000136169	SETDB2, ENSG00000092201	SUPT16H, ENSG00000134516	DOCK2, ENSG00000174100	MRPL45 
#> 	   ENSG00000174917	C19orf70, ENSG00000135365	PHF21A, ENSG00000272831	RP11-792A8.4, ENSG00000169905	TOR1AIP2, ENSG00000164970	FAM219A, ENSG00000161204	ABCF3, ENSG00000126457	PRMT1, ENSG00000146828	SLC12A9, ENSG00000168958	MFF, ENSG00000072518	MARK2 
#> 	   ENSG00000105341	ATP5SL, ENSG00000069998	CECR5, ENSG00000054148	PHPT1, ENSG00000106683	LIMK1, ENSG00000162086	ZNF75A, ENSG00000219545	RPA3-AS1, ENSG00000173442	EHBP1L1, ENSG00000072364	AFF4, ENSG00000123240	OPTN, ENSG00000108349	CASC3 
#> Negative:  ENSG00000164305	CASP3, ENSG00000120699	EXOSC8, ENSG00000162496	DHRS3, ENSG00000150776	C11orf57, ENSG00000116237	ICMT, ENSG00000013288	MAN2B2, ENSG00000163162	RNF149, ENSG00000007080	CCDC124, ENSG00000174456	C12orf76, ENSG00000131378	RFTN1 
#> 	   ENSG00000136819	C9orf78, ENSG00000099849	RASSF7, ENSG00000158805	ZNF276, ENSG00000164031	DNAJB14, ENSG00000184277	TM2D3, ENSG00000168101	NUDT16L1, ENSG00000139668	WDFY2, ENSG00000085563	ABCB1, ENSG00000060339	CCAR1, ENSG00000241839	PLEKHO2 
#> 	   ENSG00000105088	OLFM2, ENSG00000063177	RPL18, ENSG00000082701	GSK3B, ENSG00000257303	RP11-977G19.11, ENSG00000228857	AC104653.1, ENSG00000106603	COA1, ENSG00000106245	BUD31, ENSG00000082805	ERC1, ENSG00000213741	RPS29, ENSG00000226742	HSBP1L1 
#> PC_ 5 
#> Positive:  ENSG00000013288	MAN2B2, ENSG00000164051	CCDC51, ENSG00000020633	RUNX3, ENSG00000154359	LONRF1, ENSG00000164031	DNAJB14, ENSG00000164327	RICTOR, ENSG00000148218	ALAD, ENSG00000196642	RABL6, ENSG00000197223	C1D, ENSG00000272606	RP11-554J4.1 
#> 	   ENSG00000130770	ATPIF1, ENSG00000148248	SURF4, ENSG00000085788	DDHD2, ENSG00000160685	ZBTB7B, ENSG00000099330	OCEL1, ENSG00000173457	PPP1R14B, ENSG00000113282	CLINT1, ENSG00000189227	C15orf61, ENSG00000131378	RFTN1, ENSG00000187713	TMEM203 
#> 	   ENSG00000164970	FAM219A, ENSG00000198715	C1orf85, ENSG00000106683	LIMK1, ENSG00000166979	EVA1C, ENSG00000181847	TIGIT, ENSG00000168209	DDIT4, ENSG00000106367	AP1S1, ENSG00000139668	WDFY2, ENSG00000047932	GOPC, ENSG00000118518	RNF146 
#> Negative:  ENSG00000113391	FAM172A, ENSG00000166133	RPUSD2, ENSG00000241370	RPP21, ENSG00000138138	ATAD1, ENSG00000167522	ANKRD11, ENSG00000132549	VPS13B, ENSG00000142856	ITGB3BP, ENSG00000161204	ABCF3, ENSG00000170540	ARL6IP1, ENSG00000100348	TXN2 
#> 	   ENSG00000156170	NDUFAF6, ENSG00000272269	RP11-500C11.3, ENSG00000075089	ACTR6, ENSG00000126749	EMG1, ENSG00000164733	CTSB, ENSG00000161526	SAP30BP, ENSG00000111490	TBC1D30, ENSG00000204482	LST1, ENSG00000173933	RBM4, ENSG00000153140	CETN3 
#> 	   ENSG00000030582	GRN, ENSG00000109787	KLF3, ENSG00000141002	TCF25, ENSG00000179988	PSTK, ENSG00000099991	CABIN1, ENSG00000270170	NCBP2-AS2, ENSG00000169499	PLEKHA2, ENSG00000095059	DHPS, ENSG00000116497	S100PBP, ENSG00000197841	ZNF181 
#> Computing nearest neighbor graph
#> Computing SNN
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 328
#> Number of edges: 24346
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.5000
#> Number of communities: 1
#> Elapsed time: 0 seconds